% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_CausalMediation.R
\name{AA_Mediation}
\alias{AA_Mediation}
\title{Causal Mediation Analysis (CMA)}
\usage{
AA_Mediation(dataset=Metadata, XVar="Lysine", YVar="Tryptophan", AdjVar=c("Age", "Gender"), Package="mediation")
}
\arguments{
\item{dataset, }{Matrix; (Required) Metadata.}

\item{XVar, }{Character; independent variable.}

\item{YVar, }{Character; dependent variable.}

\item{MVar, }{Character; mediator.}

\item{AdjVar, }{Character; adjust variable(default: AdjVar=NULL).}

\item{Package, }{Character; (Required) package for CMA (default: Package="mediation").}
}
\value{
a list of results
  CMA model
  Sensitivity analysis of CMA model
  Table of CMA model
}
\description{
CMA examines the intermediate process(Mediation variable:M) by which the independent variable(X) affects the dependent variable(Y).
 Firstly, building Mediator Model(model.m): f(M | X + AdjVar); and then Outcome Model: f(Y = X + M + AdjVar).
 Secondly, m.out <- mediate{Y, M, X, AdjVar, model.y, model.m} to perform mediation analysis.
 Finally,
}
\details{
12/5/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSet_species)
library(Biobase)
Metadata <- pData(ExprSet_species)

# mediation package & one mediation
Mediation_res1 <- AA_Mediation(dataset=Metadata, XVar="Lysine", YVar="Tryptophan", AdjVar=c("Age", "Gender"), MVar="BMI", Package="mediation")

# intmed package or two mediation
Mediation_res2 <- AA_Mediation(dataset=Metadata, XVar="Lysine", YVar="Tryptophan", AdjVar=c("Age", "Gender"), MVar=c("BMI", "Age"), Package="intmed")

}

}
\author{
Hua Zou
}
