% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_PartialCorrelation.R
\name{AA_PartialCC}
\alias{AA_PartialCC}
\title{Association Analysis by Partial Correlation Coefficient}
\usage{
AA_PartialCC(dataset=ExpressionSet, Group_info="Group", Group_name=NULL, AdjVar=c("Age", "Gender"), Method="pcor")
}
\arguments{
\item{Group_info, }{Character; design factor(default: "Group").}

\item{Group_name, }{Character; the group for filtering(default: NULL).}

\item{AdjVar, }{Character; (Required) the variables for correlation(default: AdjVar=c("Age", "Gender")).}

\item{Method, }{Character; (Required) the method for correlation(default: Method="pcor").}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list of results
  Partial Correlation matrix(estimate, pvalue, statistics)
  Partial Correlation Coefficient table
}
\description{
The correlation between two variables (X, Y) while controling a third variable(Z).
 Here, we regard each Feature as X and each Metabolite as Y, the Age/Gender variable as Z.
}
\details{
12/4/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSetRawRB)

PartialCC_res <- AA_PartialCC(dataset=ExprSetRawRB, Group_info="Group", Group_name=NULL, AdjVar=c("Age", "Gender"), Method="pcor")

}

}
\author{
Hua Zou
}
