% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA_Spearman.R
\name{AA_Spearman}
\alias{AA_Spearman}
\title{Association Analysis by Spearman correlation coefficient}
\usage{
AA_Spearman(dataset=ExpressionSet, Group_info="Group", Group_name=NULL, Object="Feature", Pvalue=0.05)
}
\arguments{
\item{Group_info, }{Character; design factor(default: "Group").}

\item{Group_name, }{Character; the group for filtering(default: NULL).}

\item{Object, }{Character; (Required) the variables for correlation(Feature;Meta;Mix).}

\item{Pvalue, }{Numeric; significant level(default: 0.05).}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
significant difference with enriched directors
}
\description{
The correlation among features is calculated by Spearman.
}
\details{
12/3/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSet_species)

Spearman_res <- AA_Spearman(dataset=ExprSet_species, Group_info="Group", Group_name=NULL, Object="Feature", Pvalue=0.05)
Spearman_res$res
}

}
\author{
Hua Zou
}
