% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anderson.rubin.R
\name{anderson.rubin}
\alias{anderson.rubin}
\title{Anderson-Rubin test}
\usage{
anderson.rubin(Y, X, Z, b, a)
}
\arguments{
\item{Y}{The outcome variables}

\item{X}{The endogenous variables}

\item{Z}{The instruments}

\item{b}{The hypothesized value}

\item{a}{The level of the test}
}
\description{
Calculates an Anderson-Rubin test statistic and computes the test for a
hypothesized value.
}
