% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15.parametric_test.R
\name{ANOVA_one}
\alias{ANOVA_one}
\title{The one-way analysis of variance (ANOVA)}
\usage{
result <- ANOVA_one(phen, prof, "SampleID", "Stage")
}
\arguments{
\item{DNAID}{names of sampleID to connect phen and prof}

\item{GROUP}{names of group information}

\item{phen}{phenotype with sampleID and group; sampleID connected to profile.}

\item{prof}{profile table rownames->taxonomy; colnames->sampleID}
}
\value{
Returns a result of the one-way analysis of variance (ANOVA)
  type:       kind of data
  Num:        number of group
  P-value:    P by one-way analysis of variance (ANOVA)
              post test pvalue
  FDR:        adjusted by BH
  mean+/-sd   each group
  median:     each group
}
\description{
one-factor ANOVA, is an extension of independent two-samples t-test for comparing means in a situation
where there are more than two groups. In one-way ANOVA, the data is organized into several groups base on
one single grouping variable
}
\details{
07/18/2019
}
\author{
Hua Zou
}
