% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15.parametric_test.R
\name{ANOVA_two}
\alias{ANOVA_two}
\title{Two-Way ANOVA Test}
\usage{
result <- ANOVA_one(phen, prof, "SampleID", "Stage", "Group")
}
\arguments{
\item{DNAID}{names of sampleID to connect phen and prof}

\item{GROUP1}{names of group information: factor1}

\item{GROUP2}{names of group information: factor2}

\item{phen}{phenotype with sampleID and group; sampleID connected to profile.}

\item{prof}{profile table rownames->taxonomy; colnames->sampleID}
}
\value{
Returns a result of Two-Way ANOVA Test
  type:       kind of data
  Num:        number of group
  P-value:    P by Two-Way ANOVA Test
              post test pvalue
  FDR:        adjusted by BH
  mean+/-sd   each group
  median:     each group
}
\description{
Two-way ANOVA test is used to evaluate simultaneously the effect of
two grouping variables (A and B) on a response variable
}
\details{
07/18/2019
}
\author{
Hua Zou
}
