% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Assignment.R
\docType{methods}
\name{feature_are_rows<-}
\alias{feature_are_rows<-}
\alias{assign-feature_are_rows}
\alias{feature_are_rows<-,Profile_table,logical-method}
\alias{feature_are_rows<-,MyDataSet,logical-method}
\title{Manually change feature_are_rows through assignment.}
\usage{
feature_are_rows(x) <- value

\S4method{feature_are_rows}{Profile_table,logical}(x) <- value

\S4method{feature_are_rows}{MyDataSet,logical}(x) <- value
}
\arguments{
\item{x}{\code{\link{Profile_table-class}} or \code{\link{MyDataSet-class}}}

\item{value}{A logical of length equal to 1. If \code{length(value) > 1},
the additional elements will be ignored. Only the first element is assigned
to the feature_are_rows slot.}
}
\description{
The feature_are_rows slot is a logical indicating the orientation of the
abundance table contained in object \code{x}.
}
\examples{

}
