% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Assignment.R
\docType{methods}
\name{feature_names<-}
\alias{feature_names<-}
\alias{assign-feature_names}
\alias{feature_names<-,ANY,ANY-method}
\alias{feature_names<-,ANY,character-method}
\alias{feature_names<-,Profile_table,character-method}
\alias{feature_names<-,Feature_table,character-method}
\alias{feature_names<-,MyDataSet,character-method}
\title{Replace Feature identifier names}
\usage{
feature_names(x) <- value

\S4method{feature_names}{ANY,ANY}(x) <- value

\S4method{feature_names}{ANY,character}(x) <- value

\S4method{feature_names}{Profile_table,character}(x) <- value

\S4method{feature_names}{Feature_table,character}(x) <- value

\S4method{feature_names}{MyDataSet,character}(x) <- value
}
\arguments{
\item{x}{(Required). An object defined by the \code{\link{MyDataSet-package}}
that describes Features in some way.}

\item{value}{(Required). A character vector
to replace the current \code{\link{feature_names}}.}
}
\description{
Replace Feature identifier names
}
\examples{

}
