% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Assignment.R
\name{Feature_table<-}
\alias{Feature_table<-}
\alias{assign-Feature_table}
\alias{Feature_table<-,MyDataSet,Feature_table-method}
\alias{Feature_table<-,MyDataSet,ANY-method}
\alias{Feature_table<-,Feature_table,Feature_table-method}
\alias{Feature_table<-,Feature_table,ANY-method}
\title{Assign a (new) Feature Table to \code{x}}
\usage{
Feature_table(x) <- value

\S4method{Feature_table}{MyDataSet,Feature_table}(x) <- value

\S4method{Feature_table}{MyDataSet,ANY}(x) <- value

\S4method{Feature_table}{Feature_table,Feature_table}(x) <- value

\S4method{Feature_table}{Feature_table,ANY}(x) <- value
}
\arguments{
\item{x}{(Required). \code{\link{MyDataSet-class}}}

\item{value}{(Required). \code{\link{Feature_table-class}}.}
}
\description{
Assign a (new) Feature Table to \code{x}
}
\examples{

}
