% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Assignment.R
\docType{methods}
\name{Profile_table<-}
\alias{Profile_table<-}
\alias{assign-Profile_table}
\alias{Profile_table<-,MyDataSet,Profile_table-method}
\alias{Profile_table<-,Profile_table,Profile_table-method}
\alias{Profile_table<-,MyDataSet,MyDataSet-method}
\title{Assign a new Profile Table to \code{x}}
\usage{
Profile_table(x) <- value

\S4method{Profile_table}{MyDataSet,Profile_table}(x) <- value

\S4method{Profile_table}{Profile_table,Profile_table}(x) <- value

\S4method{Profile_table}{MyDataSet,MyDataSet}(x) <- value
}
\arguments{
\item{x}{(Required). \code{\link{MyDataSet-class}}}

\item{value}{(Required).
\code{\link{Profile_table-class}}
or
\code{\link{MyDataSet-class}}.}
}
\description{
Assign a new Profile Table to \code{x}
}
\examples{

}
