% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Assignment.R
\name{Sample_data<-}
\alias{Sample_data<-}
\alias{assign-Sample_data}
\title{Assign (new) Sample_data to \code{x}}
\usage{
Sample_data(x) <- value
}
\arguments{
\item{x}{(Required). \code{\link{MyDataSet-class}}. The object to modify.}

\item{value}{(Required). Either a \code{\link{Sample_data-class}},
a \code{data.frame} that can be coerced into \code{\link{Sample_data-class}},
or a \code{\link{MyDataSet-class}} that contains a
suitable \code{Sample_data} component to assign to \code{x}. If unsure,
try \code{\link{Sample_data}}\code{(value)}, which should return a
\code{\link{Sample_data-class}} object without error.}
}
\value{
No return. This is an assignment statement.
}
\description{
This replaces the current \code{Sample_data} component of \code{x} with
\code{value}, if \code{value} is a \code{\link{Sample_data-class}}. However,
if \code{value} is a \code{data.frame}, then \code{value} is first coerced to
a \code{\link{Sample_data-class}}, and then assigned. Alternatively, if
\code{value} is \code{\link{MyDataSet-class}}, then the
\code{\link{Sample_data}} component will first be accessed from \code{value}
 and then assigned. This makes possible some concise assignment/replacement
 statements when adjusting, modifying, or building subsets of
 experiment-level data. See some examples below.
}
\details{
Internally, this re-builds the \code{\link{MyDataSet-class}} object using
the standard \code{\link{MyDataSet}} constructor. Thus, index mismatches
between sample-describing components will not be allowed, and subsetting
will occurr automatically such that only the intersection of sample IDs
are included in any components. This has the added benefit of re-checking
(internally) for any other issues.
}
\examples{

}
