% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Assignment.R
\docType{methods}
\name{sample_names<-}
\alias{sample_names<-}
\alias{assign-sample_names}
\alias{sample_names<-,ANY,ANY-method}
\alias{sample_names<-,ANY,character-method}
\alias{sample_names<-,Profile_table,character-method}
\alias{sample_names<-,Sample_data,character-method}
\alias{sample_names<-,sample_data,character-method}
\alias{sample_names<-,MyDataSet,character-method}
\title{Replace Sample identifier names}
\usage{
sample_names(x) <- value

\S4method{sample_names}{ANY,ANY}(x) <- value

\S4method{sample_names}{ANY,character}(x) <- value

\S4method{sample_names}{Profile_table,character}(x) <- value

\S4method{sample_names}{Sample_data,character}(x) <- value

\S4method{sample_names}{MyDataSet,character}(x) <- value
}
\arguments{
\item{x}{(Required). An object defined by the \code{\link{MyDataSet-package}}
that describes Samples in some way.}

\item{value}{(Required). A character vector
to replace the current \code{\link{sample_names}}.}
}
\description{
Replace Sample identifier names
}
\examples{

}
