% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifyFilesWithKeyWrodsInFileName.R
\name{BatchModifyFileExtended}
\alias{BatchModifyFileExtended}
\title{Extended Batch-Modify-File Function}
\usage{
BatchModifyFileExtended(
  Dir,
  FileNameKeyWords,
  FileNamepattern = NULL,
  FileName.ignore.case = FALSE,
  KeyWordOldLocation,
  Location.ignore.case = FALSE,
  keyWordOld,
  keyWordNew,
  Replace = FALSE,
  SaveOld = TRUE
)
}
\arguments{
\item{Dir}{The directory where you target files exist.}

\item{FileNameKeyWords}{key words for filenames.}

\item{FileNamepattern}{filename pattern, such as "*.Rmd",defaul is NULL.}

\item{FileName.ignore.case}{whether to ignore case of filename key words for searching files. default is FALSE.}

\item{KeyWordOldLocation}{key words used to find target line, could be a character vector.}

\item{Location.ignore.case}{whether to ignore case of line key words for searching files. default is FALSE.}

\item{keyWordOld}{The old word needs to be replaced.}

\item{keyWordNew}{The new word.}

\item{Replace}{Whether to replace the old file. Default is FALSE.}

\item{SaveOld}{Wheter to save the old file, this will suffix date to the file name.}
}
\value{

}
\description{
More pararmeters included in this function, Details see the Examples!
}
\examples{
# Better not use "../../" or "./" in Dir! for it probably will cause mistakes!
BatchModifyFileExtended("./man/",
                        FileNameKeyWords = c("a","m"),
                        FileNamepattern = NULL,
                        FileName.ignore.case = TRUE,
                        KeyWordOldLocation = c("a","b"),
                        Location.ignore.case = FALSE,
                        keyWordOld = "roxygen2",
                        keyWordNew = "####",
                        Replace = FALSE,
                        SaveOld = TRUE)
}
