% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein.poly.R
\name{bernstein.poly}
\alias{bernstein.poly}
\title{Bernstein polynomial estimation}
\usage{
bernstein.poly(Y, X, K = 1, space)
}
\arguments{
\item{Y}{A vector of outcome variables}

\item{X}{A vector of predictors}

\item{K}{The order of the polynomial}

\item{space}{The values at which to predict}
}
\description{
This performs Bernstein polynomial estimation on a given space
}
