% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(dt, func, S)
}
\arguments{
\item{dt}{The data to be used in the given function}

\item{func}{The function to be repeated.  The output of func should be the
estimate of interest.}

\item{S}{The number of samples to draw}
}
\description{
This performs a bootstrap for a given function
}
