% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.visualization.R
\name{boxplot2}
\alias{boxplot2}
\title{boxplot2}
\usage{
boxplot2(dat = DATA)
}
\arguments{
\item{DATA}{Table.}
}
\value{
boxplot
}
\description{
boxplot
}
\examples{
boxplot2(DATA)

}
