% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.var.R
\name{classic.var}
\alias{classic.var}
\title{Classical homoskedastic variance estimation}
\usage{
classic.var(Y, X, beta, const = TRUE)
}
\arguments{
\item{Y}{The Y data used in estimation}

\item{X}{The X data used in estimation}

\item{beta}{The vector of linear regression estimates}

\item{const}{If TRUE, adds a constant column to the data}
}
\description{
Classical homoskedastic variance estimation
}
