% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.function.R
\name{Correlation_matrix}
\alias{Correlation_matrix}
\title{CorrMatrix}
\usage{
Correlation_matrix(dat = DATA2)
}
\arguments{
\item{DATA}{Table mtcars}
}
\value{
Matrix
}
\description{
The flattenCorrMatrix function will format the correlation matrix into a table of 4 columns:
row names, column names, the correlation coefficient between each variable and the others, and the p-values
cormat : matrix of the correlation coefficients
pmat : matrix of the correlation p-values
}
\examples{
Correlation_matrix(DATA2)

}
