% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_DESeq2.R
\name{DA_DESeq2}
\alias{DA_DESeq2}
\title{Differential Expression Analysis by DEseq2 Package}
\usage{
DA_DESeq2(dataset=ExpressionSet, Group_info="Group", Pvalue=0.05, Log2FC=1)
}
\arguments{
\item{Group_info, }{Character; design factor(default: "Group").}

\item{Group_name, }{Character; the group for comparison.}

\item{Pvalue, }{Numeric; significant level(default: 0.05).}

\item{Log2FC, }{Numeric; log2FoldChange(default: 1).}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list object:
  DESeq object
  DESeq results
  significant difference with enriched directors
}
\description{
DEseq2 requires the count data (a matrix of integer values) to input. The normalization method is to use the standard factor vector per feature.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

data(ExprSet_species_count)

DESeq2_res <- DA_DESeq2(dataset=ExprSet_species_count, Group_info="Group", Pvalue=0.05, Log2FC=1)
DESeq2_res$res

}
\author{
Hua Zou
}
