% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_edgeR.R
\name{DA_edgeR}
\alias{DA_edgeR}
\title{Differential Expression Analysis by edgeR Package}
\usage{
DA_edgeR(dataset=ExpressionSet, Group_info="Group", Group_name=c("HC", "AA"), Pvalue=0.05, Log2FC=1)
}
\arguments{
\item{Group_info, }{Character; design factor(default: "Group").}

\item{Group_name, }{Character; (Required) the group for comparison.}

\item{Pvalue, }{Numeric; significant level(default: 0.05).}

\item{Log2FC, }{Numeric; log2FoldChange(default: 1).}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list object:
  edgeR results
  significant difference with enriched directors
}
\description{
edgeR requires the count data (a matrix of integer values) to input.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSet_species_count)

edgeR_res <- DA_edgeR(dataset=ExprSet_species_count, Group_info="Group", Group_name=c("HC", "AA"), Pvalue=0.05, Log2FC=1)
edgeR_res$res
}

}
\author{
Hua Zou
}
