% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_LEfSe.R
\name{DA_LEfSe}
\alias{DA_LEfSe}
\title{Differential Expression Analysis by LEfSe Package}
\usage{
DA_LEfSe(dataset=ExpressionSet, Group_info="Group", Group_name=c("HC", "AA"), kw.p=0.05, wl.p=0.05, Lda=2)
}
\arguments{
\item{Group_info, }{Character; design factor(default: "Group").}

\item{Group_name, }{Character; (Required) the group for comparison.}

\item{kw.p, }{Numeric; significant level of kruskal.test(default: 0.05).}

\item{wl.p, }{Numeric; significant level of wilcox(default: 0.05).}

\item{Lda, }{Numeric; LDA score(default: 2).}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
significant difference with enriched directors
}
\description{
LEfSe method for microbiome biomarker discovery.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

data(ExprSet_species_count)

LEfSe_res <- DA_LEfSe(dataset=ExprSet_species_count, Group_info="Group", Group_name=c("HC", "AA"), kw.p=0.05, wl.p=0.05, Lda=2)
LEfSe_res

}
\author{
Hua Zou
}
