% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_limma_voom.R
\name{DA_Limma}
\alias{DA_Limma}
\title{Differential Expression Analysis by limma Package}
\usage{
DA_Limma(dataset=ExpressionSet, Group_info="Group", Group_name=c("HC", "AA"), Pvalue=0.05, Log2FC=1)
}
\arguments{
\item{Group_info, }{Character; design factor(default: "Group").}

\item{Group_name, }{Character; (Required) the group for comparison.}

\item{Pvalue, }{Numeric; significant level(default: 0.05).}

\item{Log2FC, }{Numeric; log2FoldChange(default: 1).}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list object:
  limma results
  significant difference with enriched directors
}
\description{
limma requires the count data (a matrix of integer values) to input. The *Voom* would transform the count data for further data analysis.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

data(ExprSet_species_count)

Limma_res <- DA_Limma(dataset=ExprSet_species_count, Group_info="Group", Group_name=c("HC", "AA"), Pvalue=0.05, Log2FC=1)
Limma_res$res

}
\author{
Hua Zou
}
