% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_WilcoxRankSumTest.R
\name{DA_Wilcox}
\alias{DA_Wilcox}
\title{Differential Expression Analysis by Wilcox-rank-sum test}
\usage{
DA_Wilcox(dataset=ExpressionSet, Group_info="Group", Group_name=c("HC", "AA"), Pvalue=0.05)
}
\arguments{
\item{Group_info, }{Character; design factor(default: "Group").}

\item{Group_name, }{Character; (Required) the group for comparison.}

\item{Pvalue, }{Numeric; significant level(default: 0.05).}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
significant difference with enriched directors
}
\description{
Wilcox-rank-sum test is non-parameter test method, and also use for the data with non-normal distribution(Gauss Distribution).
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSet_species)

Wilcox_res <- DA_Wilcox(dataset=ExprSet_species, Group_info="Group", Group_name=c("HC", "AA"), Pvalue=0.05)
Wilcox_res$res
}

}
\author{
Hua Zou
}
