% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanechnikov.R
\name{epanechnikov}
\alias{epanechnikov}
\title{Epanechnikov Kernel}
\usage{
epanechnikov(x, h)
}
\arguments{
\item{x}{The x value of the function}

\item{h}{The one-half bandwidth (i.e. x+h,x-h)}
}
\description{
The Epanechnikov kernel
}
