% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_DataAccess.R
\docType{methods}
\name{feature_are_rows}
\alias{feature_are_rows}
\alias{feature_are_rows,ANY-method}
\alias{feature_are_rows,Profile_table-method}
\alias{feature_are_rows,MyDataSet-method}
\title{Access feature_are_rows slot from Profile_table objects.}
\usage{
feature_are_rows(mydataset)

\S4method{feature_are_rows}{ANY}(mydataset)

\S4method{feature_are_rows}{Profile_table}(mydataset)

\S4method{feature_are_rows}{MyDataSet}(mydataset)
}
\arguments{
\item{mydataset}{(Required). \code{\link{MyDataSet-class}}, or \code{\link{Profile_table-class}}.}
}
\value{
A logical indicating the orientation of the Profile_table
}
\description{
Access feature_are_rows slot from Profile_table objects.
}
\seealso{
\code{\link{Profile_table}}
}
