% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_DataAccess.R
\docType{methods}
\name{feature_names}
\alias{feature_names}
\alias{feature_names,ANY-method}
\alias{feature_names,MyDataSet-method}
\alias{feature_names,Profile_table-method}
\alias{feature_names,Feature_table-method}
\alias{feature_names,Sample_data-method}
\title{Get feature names.}
\usage{
feature_names(mydataset)

\S4method{feature_names}{ANY}(mydataset)

\S4method{feature_names}{MyDataSet}(mydataset)

\S4method{feature_names}{Profile_table}(mydataset)

\S4method{feature_names}{Feature_table}(mydataset)

\S4method{feature_names}{Sample_data}(mydataset)
}
\arguments{
\item{mydataset}{\code{\link{MyDataSet-class}}, \code{\link{Profile_table-class}},
\code{\link{Feature_table-class}}}
}
\value{
A character vector of the names of the species in \code{mydataset}.
}
\description{
Get feature names.
}
\examples{

}
