% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_DataAccess.R
\name{feature_sums}
\alias{feature_sums}
\title{Returns the total number of individuals observed from each feature.}
\usage{
feature_sums(x)
}
\arguments{
\item{x}{\code{\link{Profile_table-class}}, or \code{\link{MyDataSet-class}}.}
}
\value{
A \code{\link{numeric-class}} with length equal to the number of species
 in the table, name indicated the feature ID, and value equal to the sum of
 all individuals observed for each feature in \code{x}.
}
\description{
A convenience function equivalent to rowSums or colSums, but where
the orientation of the Profile_table is automatically handled.
}
\examples{

}
