% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Classes.R
\docType{class}
\name{Feature_table-class}
\alias{Feature_table-class}
\title{An S4 class that holds feature attributions data as a character
data.frame}
\description{
Row indices represent feature, columns represent feature attributions.
}
\details{
\describe{

   \item{.Data}{data-frame data, inherited from the matrix class.}

   \item{row.names}{
     Also inherited from the matrix class;
      it should contain the sample names.
   }

   \item{names}{Inherited from the matrix class.}

 }
}
