% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filter.R
\name{filter_method}
\alias{filter_method}
\title{Filter samples or features in `Profile_table` by Occurrences}
\usage{
filter_method(Object, Cutoff=0.2, Type="Both")
}
\arguments{
\item{Cutoff, }{Numeric; the threshold for filtering (default: Cutoff=0.2).}

\item{Type, }{Character; the type of filtering data ("Both", "Features", "Samples").}

\item{Object, }{Class; a [`MyDataSet-class`].}
}
\value{
A filtered `object` with the Cutoff.
}
\description{
Filter samples or features in `Profile_table` by Occurrences,
which means the samples or features will be discard if they could not pass the cutoff.
}
\details{
12/6/2021 Guangzhou China
}
\examples{
data(mydataset)
filter_method(Object, Cutoff=0.2, Type="Both")

}
\author{
Hua Zou
}
