% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filter.R
\name{filter_profile}
\alias{filter_profile}
\title{Filter samples or features in `Profile_table` by Occurrences}
\usage{
NormalizeFun(Vector, type="Zscore")
}
\arguments{
\item{Cutoff, }{Numeric; the threshold for filtering (default: Cutoff=0.2).}

\item{Type, }{Character; the type of filtering data ("Feature" or "Sample").}

\item{Object, }{Class; a [`Profile_table-class`] or [`MyDataSet-class`].}

\item{Each, }{Logical; filtering Features each group or whole data(default: Each=FALSE).}
}
\value{
A filtered `object` with the Cutoff.
}
\description{
Filter samples or features in `Profile_table` by Occurrences,
which means the samples or features will be discard if they could not pass the cutoff.
}
\details{
12/6/2021 Guangzhou China
}
\examples{

}
\author{
Hua Zou
}
