% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifyFilesWithKeyWrodsInFileName.R
\name{FindFiles}
\alias{FindFiles}
\title{Search files with filename include specified key words}
\usage{
FindFiles(Dir, FileNameKeyWords, ignore.case, pattern)
}
\arguments{
\item{Dir}{The directory where you target files exist.}

\item{FileNameKeyWords}{key words for filenames.}

\item{ignore.case}{whether to ignore case of filename key words for searching files. default is FALSE.}

\item{pattern}{filename pattern, such as "*.Rmd",defaul is NULL.}
}
\value{

}
\description{
Search files with filename include specified key words
}
