% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.nonparametric_test.R
\name{friedman_test}
\alias{friedman_test}
\title{Friedman Test}
\usage{
friedman_test(x, y, DNAID, PID, GROUP,grp1=NULL, grp2=NULL,grp3=NULL)
}
\arguments{
\item{x}{x with sampleID and group; sampleID connected to x}

\item{y}{y table rownames->taxonomy; colnames->sampleID}

\item{DNAID}{names of sampleID to connect x and y}

\item{PID}{id for paired test}

\item{GROUP}{names of group information}

\item{grp1}{one of groups}

\item{grp2}{one of groups}

\item{grp3}{one of groups}
}
\value{
Returns a result of Friedman Test

type:       kind of data

Number:     number of group

P-value:    Pvalue by Friedman Test or  pvalue by post test

FDR:        correct by BH in friedman test

Mean+SD:    each group

Median:     each group
}
\description{
The Friedman test is a non-parametric statistical test developed by Milton Friedman.
Similar to the parametric repeated measures ANOVA, it is used to detect differences
in treatments across multiple test attempts. The procedure involves ranking each row (or block) together,
then considering the values of ranks by columns
}
\details{
07/18/2019
}
\examples{
result <- friedman_test(phen, prof, "SampleID", "ID", "Stage")
}
\author{
Hua Zou
}
