% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_Ordination.R
\name{GB_MDS}
\alias{GB_MDS}
\title{Dimensionality Reduction Analysis: Multidimensional Scaling(MDS)}
\usage{
GB_MDS(dataset=ExpressionSet, Method="Classic", Group_info="Group")
}
\arguments{
\item{Method, }{Character; method for MDS (default: Classic).}

\item{Group_info, }{Character; the group for plot(default: "Group").}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list object:
  MDS score
  Result of PERMANOVA
}
\description{
MDS use the distance among samples, which calculates through the multiple variables.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSet_species)

MDS_res <- GB_MDS(dataset=ExprSet_species, Method="Classic", Group_info="Group")
MDS_res$MDS
}

}
\author{
Hua Zou
}
