% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_Ordination.R
\name{GB_PCoA}
\alias{GB_PCoA}
\title{Dimensionality Reduction Analysis: Principal Coordinate Analysis(PCoA)}
\usage{
GB_PCoA(dataset=ExpressionSet, Method="bray", Group_info="Group")
}
\arguments{
\item{Method, }{Character; method to calculate the distance(default: bray).}

\item{Group_info, }{Character; the group for plot(default: "Group").}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list object:
  PCoA score
  Result of PERMANOVA
}
\description{
PCoA uses the distance among samples, which calculates through the multiple variables.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSet_species)

PCoA_res <- GB_PCoA(dataset=ExprSet_species, Method="bray", Group_info="Group")
PCoA_res$PCoA
}

}
\author{
Hua Zou
}
