% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_Ordination.R
\name{GB_tsne}
\alias{GB_tsne}
\title{Dimensionality Reduction Analysis: t-Distributed Stochastic Neighbor Embedding(t-SNE)}
\usage{
GB_tsne(dataset=ExpressionSet, Perplexity=20, Group_info="Group")
}
\arguments{
\item{Perplexity, }{Numeric; (Required) numeric; Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation).}

\item{Group_info, }{Character; the group for plot(default: "Group").}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list object:
  t-SNE score
  Result of PERMANOVA
}
\description{
Visualization of High Dimensional Data using t-SNE with R.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSet_species)

tsne_res <- GB_tsne(dataset=ExprSet_species, Perplexity=20, Group_info="Group")
tsne_res$tsne
}

}
\author{
Hua Zou
}
