% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_Ordination.R
\name{GB_umap}
\alias{GB_umap}
\title{Dimensionality Reduction Analysis: Uniform Manifold Approximation and Projection (UMAP)}
\usage{
GB_umap(dataset=ExpressionSet, Group_info="Group")
}
\arguments{
\item{Group_info, }{Character; the group for plot(default: "Group").}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list object:
  umap score
  Result of PERMANOVA
}
\description{
UMAP: a non-linear dimensionality reduction algorithm.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSet_species)

umap_res <- GB_umap(dataset=ExprSet_species, Group_info="Group")
umap_res$umap
}

}
\author{
Hua Zou
}
