% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_MyDataSet.R
\name{get.component.classes}
\alias{get.component.classes}
\title{Show the component objects classes and slot names.}
\usage{
get.component.classes()
}
\value{
a character vector of the component objects classes, where each
element is named by the corresponding slot name in the MyDataSet-class.
}
\description{
There are no arguments to this function. It returns a named character
when called, which can then be used for tests of component data types, etc.
}
\examples{
#

}
\keyword{internal}
