% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_imputedExprSet}
\alias{get_imputedExprSet}
\title{the new imputed ExpressionSet object}
\usage{
get_imputedExprSet(dataset=ExpressionSet)
}
\arguments{
\item{dataset, }{ExpressionSet; (Required) imputed ExpressionSet object.}
}
\value{
an imputed ExpressionSet object
}
\description{
After imputing features, we need to obtain the new ExpressionSet object.
}
\details{
12/13/2021 Guangzhou China
}
\examples{

\donttest{

}

}
\author{
Hua Zou
}
