% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_NormalizedExprSet}
\alias{get_NormalizedExprSet}
\title{the new Normalized ExpressionSet object}
\usage{
get_NormalizedExprSet(dataset=ExpressionSet, expr=tranformedMatrix)
}
\arguments{
\item{dataset, }{ExpressionSet; (Required) old ExpressionSet object.}

\item{expr, }{matrix; (Required) the transformed expression matrix.}
}
\value{
an Normalized ExpressionSet object
}
\description{
After Normalizing the expression, we need to obtain the new Normalized ExpressionSet object.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{

}

}
\author{
Hua Zou
}
