% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_TransformedExprSet}
\alias{get_TransformedExprSet}
\title{the new transformed ExpressionSet object}
\usage{
get_TransformedExprSet(dataset=ExpressionSet, expr=tranformedMatrix)
}
\arguments{
\item{dataset, }{ExpressionSet; (Required) old ExpressionSet object.}

\item{expr, }{matrix; (Required) the transformed expression matrix.}
}
\value{
an transformed ExpressionSet object
}
\description{
After tranformed the expression, we need to obtain the new transformed ExpressionSet object.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{

}

}
\author{
Hua Zou
}
