% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_TrimedExprSet}
\alias{get_TrimedExprSet}
\title{the new filtered ExpressionSet object}
\usage{
get_TrimedExprSet(dataset=ExpressionSet)
}
\arguments{
\item{dataset, }{ExpressionSet; (Required) unbalanced ExpressionSet object.}
}
\value{
an filtered ExpressionSet object
}
\description{
After filtering samples or features, we need to obtain the new ExpressionSet object.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{

}

}
\author{
Hua Zou
}
