% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.var.R, R/hc0.var.R
\name{hc0.var}
\alias{hc0.var}
\title{Classical homoskedastic variance estimation}
\usage{
hc0.var(Y, X, beta, const = TRUE)

hc0.var(Y, X, beta, const = TRUE)
}
\arguments{
\item{Y}{The Y data used in estimation}

\item{X}{The X data used in estimation}

\item{beta}{The vector of linear regression estimates}

\item{const}{If TRUE, adds a constant column to the data}
}
\description{
Computes HC0 covariance matrix of estimates
}
