% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc1.var.R
\name{hc1.var}
\alias{hc1.var}
\title{Heteroskedasticity robust HC1 covariance matrix}
\usage{
hc1.var(Y, X, beta, const = TRUE)
}
\arguments{
\item{Y}{The Y data used in estimation}

\item{X}{The X data used in estimation}

\item{beta}{The vector of linear regression estimates}

\item{const}{If TRUE, adds a constant column to the data}
}
\description{
Computes HC1 covariance matrix of estimates
}
