% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{Creates a heatmap plot fork by raivokolde/pheatmap wrap and make it more simple}
\usage{
heatmap_plot(data, Scale = "column", showRowNames = F,
  showColNames = TRUE, borderColor = "grey", clusterRows = TRUE,
  clusterCols = TRUE, clusterDistanceRows = "euclidean",
  clusterDistanceCols = "euclidean", clusterMethod = "complete",
  treeHeightRow = 0, clusterLegend = TRUE, Main = NA, fontSize = 10,
  fontSizeRow = fontSize, fontSizeCol = fontSize, clusterGroup = NA,
  Color = colorRampPalette(rev(RColorBrewer::brewer.pal(n = 7, name =
  "RdYlGn")))(100), saveType = c("both", "pdf", "png"), fileName = NULL,
  filePath = "", width = 8, height = 6)
}
\arguments{
\item{data}{numeric matrix of the values to be plotted.}

\item{Scale}{character indicating if the values should be centered and scaled in
either the row direction or the column direction, or none. Corresponding values are
\code{"row"}, \code{"column"} and \code{"none"} default is \code{"column"}}

\item{borderColor}{color of cell borders on heatmap, use \code{grey} if no border should be drawn.}

\item{clusterRows}{boolean values determining if rows should be clustered or \code{hclust} object,}

\item{clusterCols}{boolean values determining if columns should be clustered or \code{hclust} object.}

\item{Main}{the title of the plot}

\item{fontSize}{base fontsize for the plot}

\item{fontSizeRow}{fontsize for rownames (Default: fontsize)}

\item{fontSizeCol}{fontsize for colnames (Default: fontsize)}

\item{clusterGroup}{a list of character to divide your group,default is \code{NA}.}

\item{fileName}{file name. Filetype is decided by
the extension in the path. Currently following formats are supported: png, pdf, tiff,
 bmp, jpeg. Even if the plot does not fit into the plotting window, the file size is
calculated so that the plot would fit there, unless specified otherwise.}

\item{filePath}{file path where to save the picture.}

\item{color}{vector of colors used in heatmap.}

\item{clusteringDistanceRows}{distance measure used in clustering rows. Possible
values are \code{"correlation"} for Pearson correlation and all the distances
supported by \code{\link{dist}}, such as \code{"euclidean"}, etc. If the value is none
of the above it is assumed that a distance matrix is provided.}

\item{clusteringDistanceCols}{distance measure used in clustering columns. Possible
values the same as for clustering_distance_rows.}

\item{clusteringMethod}{clustering method used. Accepts the same values as
\code{\link{hclust}}.}

\item{treeHeightCol}{the height of a tree for columns, if these are clustered.
Default value 50 points.}

\item{legend}{logical to determine if legend should be drawn or not.}

\item{Width}{manual option for determining the output file width in inches.}

\item{Height}{manual option for determining the output file height in inches.}
}
\value{
A heatmap plot file of pdf or png.
}
\description{
The function also allows to aggregate the rows using kmeans clustering. This is
advisable if number of rows is so big that R cannot handle their hierarchical
}
