% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_imputation.R
\name{impute_GlobalStandard}
\alias{impute_GlobalStandard}
\alias{impute_knn}
\title{Global Standard imputation}
\usage{
impute_GlobalStandard(object, width = 0.3, downshift = 1.8)

impute_knn(object)
}
\arguments{
\item{object, }{Object; a [`matrix`](Row->Features; Column->Samples).}
}
\description{
Global Standard imputation is used the Global Median and standard as the cutoff to impute the missing value.
We calculate the mean or median for all non missing values of that variable then replace missing value with mean or median
}
