% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_MyDataSet.R
\docType{methods}
\name{index_reorder}
\alias{index_reorder}
\alias{index_reorder,MyDataSet-method}
\title{Force index order of MyDataSet objects}
\usage{
index_reorder(ps, index_type)

\S4method{index_reorder}{MyDataSet}(ps, index_type = "both")
}
\arguments{
\item{ps}{(Required). A \code{\link{MyDataSet-class}} instance.}

\item{index_type}{(Optional). A character string
specifying the indices to properly order.
Supported values are \code{c("both", "Features", "Samples")}.
Default is \code{"both"}, meaning samples and Features indices
will be checked/re-ordered.}
}
\description{
Force index order of MyDataSet objects
}
\examples{
## data("GlobalPatterns")
## GP = index_reorder(GlobalPatterns)
}
\keyword{internal}
