% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.function.R
\name{install_suggest_packages}
\alias{install_suggest_packages}
\title{Install suggest packages}
\usage{
install_suggest_packages(packages_name_list = c("randomForest"))
}
\arguments{
\item{packages}{a list of packages}
}
\description{
The install_suggest_packages function will install the suggest packages
}
\examples{
install_suggest_packages(packages_name_list)

}
