% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_installation.R
\name{InstallPackage}
\alias{InstallPackage}
\title{Installation of R packages}
\usage{
InstallPackage(Package=Packages, Type="CRAN")
}
\arguments{
\item{Package, }{Character; (Required).}

\item{Type, }{Character; the kind of repository(default: "CRAN").}
}
\value{

}
\description{
Installing the R packages coming from CRAN and bioconductor.
}
\details{
12/4/2021 Guangzhou China
}
\examples{

\donttest{
   Packages_list <- c("DESeq2", "GSVA")
   InstallPackage(Package=Packages_list, Type="bioconductor")
}

}
\author{
Hua Zou
}
