% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{intersect_feature}
\alias{intersect_feature}
\title{Returns the intersection of species and samples for the components of x}
\usage{
intersect_feature(x)
}
\arguments{
\item{x}{(Required). A \code{\link{ExpressionSet-class}} object
that contains 2 or more components
that in-turn describe feature.}
}
\value{
Returns a character vector of only those feature that are present in
 all feature-describing components of \code{x}.
}
\description{
This function is used internally as part of the infrastructure to ensure that
component data types in a ExpressionSet-object have exactly the same feature.
It relies heavily on the \code{\link{Reduce}} function to determine the
strictly common species.
}
\examples{

}
\seealso{
\code{\link{Reduce}}, \code{\link{intersect}}
}
\keyword{internal}
