% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jive.R
\name{jive}
\alias{jive}
\title{Jackknife IV estimator}
\usage{
jive(Y, X, Z, W)
}
\arguments{
\item{Y}{The outcome variables}

\item{X}{The endogenous variables}

\item{Z}{The excluded instruments}

\item{W}{The included instruments or controls}
}
\description{
Estimates jackknife IV coefficients.
}
