% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jive.var.R
\name{jive.var}
\alias{jive.var}
\title{Jackknife IV covariance matrix}
\usage{
jive.var(Y, X, Z, W, beta)
}
\arguments{
\item{Y}{The Y data used in estimation}

\item{X}{The X data used in estimation}

\item{Z}{The Z used in estimation}

\item{W}{The W used in estimation}

\item{beta}{The vector of regression estimates}
}
\description{
Computes jackknife IV covariance matrix of estimates
}
