% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.nonparametric_test.R
\name{kruskal_test}
\alias{kruskal_test}
\title{Kruskal Test}
\usage{
result <- kruskal_test(phen, prof, "SampleID", "ID", "Stage")
}
\arguments{
\item{DNAID}{names of sampleID to connect phen and prof}

\item{GROUP}{names of group information}

\item{grp1}{one of groups}

\item{grp2}{one of groups}

\item{grp3}{one of groups}

\item{phen}{phenotype with sampleID and group; sampleID connected to profile.}

\item{prof}{profile table rownames->taxonomy; colnames->sampleID}
}
\value{
Returns a result of Kruskal Test
  type:       kind of data
  Number:     number of group
  P-value:    Pvalue by Friedman Test
              pvalue by post test
  Mean+SD:    each group
  Median:     each group
}
\description{
The Kruskal–Wallis test by ranks, Kruskal–Wallis H test (named after William Kruskal and W. Allen Wallis),
or one-way ANOVA on ranks is a non-parametric method for testing whether samples originate from the same distribution.
It is used for comparing two or more independent samples of equal or different sample sizes.
 It extends the Mann–Whitney U test, which is used for comparing only two groups
}
\details{
07/18/2019
}
\author{
Hua Zou
}
