% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.spline.R
\name{lin.spline}
\alias{lin.spline}
\title{Linear Spline}
\usage{
lin.spline(Y, X, knots = 4, space)
}
\arguments{
\item{Y}{A vector of outcomes}

\item{X}{A vector of predictors}

\item{knots}{The number of knots in the spline}

\item{space}{A vector of X values at which to predict}
}
\description{
This function fits a linear spline of Y on X and returns predicted values
}
