% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.visualization.R
\name{linechart1}
\alias{linechart1}
\title{linechart1}
\usage{
linechart1(dat = DATA)
}
\arguments{
\item{DATA}{Table}
}
\value{
linechart
}
\description{
linechart
}
\examples{
linechart1(DATA)

}
