% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.polynomial.R
\name{local.polynomial}
\alias{local.polynomial}
\title{Local polynomial regression}
\usage{
local.polynomial(Y, X, h, order = 0, space)
}
\arguments{
\item{Y}{A vector of outcomes}

\item{X}{A vector of predictors}

\item{h}{The one-half bandwidth to use in the Epanechnikov kernel}

\item{order}{The order of the polynomial}

\item{space}{The values at which to predict}
}
\description{
This performs local polynomial regression in one dimension
}
