% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit.R
\name{loglik}
\alias{loglik}
\title{Probit estimation}
\usage{
loglik(beta, Y, X)
}
\arguments{
\item{Y}{The outcome variables}

\item{X}{The explanatory variables
NOTE: Add variance later}
}
\description{
Estimates probit parameters and fitted values
}
