% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Creates a manhattan plot fork by qqman/manhattan wrap and make it more simple}
\usage{
manhattan_plot(data, Chr = "CHR", Bp = "BP", P = "p", Snp = "SNP",
  AnnotatePval = NULL, AnnotateTop = TRUE, Col = NULL, ChrLabs = NULL,
  SuggestiveLine = -log10(5e-05), GenomeWideLine = -log10(5e-08),
  Highlight = NULL, Logp = TRUE, fileName = NULL, filePath = "",
  width = 8, height = 6, saveType = c("both", "png", "pdf"), ...)
}
\arguments{
\item{data}{A data.frame with columns "BP," "CHR," "P," and optionally, "SNP."}

\item{Chr}{A string denoting the column name for the chromosome. Defaults to
PLINK's "CHR." Said column must be numeric. If you have X, Y, or MT
chromosomes, be sure to renumber these 23, 24, 25, etc.}

\item{Bp}{A string denoting the column name for the chromosomal position.
Defaults to PLINK's "BP." Said column must be numeric.}

\item{P}{A string denoting the column name for the p-value. Defaults to
PLINK's "P." Said column must be numeric.}

\item{Snp}{A string denoting the column name for the SNP name (rs number).
Defaults to PLINK's "SNP." Said column should be a character.}

\item{AnnotatePval}{If set, SNPs below this p-value will be annotated on the plot.}

\item{AnnotateTop}{If TRUE, only annotates the top hit on each chromosome that is below the annotatePval threshold.}

\item{Col}{A character vector indicating which colors to alternate.}

\item{ChrLabs}{A character vector equal to the number of chromosomes
specifying the chromosome labels (e.g., \code{c(1:22, "X", "Y", "MT")}).}

\item{SuggestiveLine}{Where to draw a "suggestive" line. Default
-log10(1e-5). Set to FALSE to disable.}

\item{Logp}{If TRUE, the -log10 of the p-value is plotted. It isn't very
useful to plot raw p-values, but plotting the raw value could be useful for
other genome-wide plots, for example, peak heights, bayes factors, test
statistics, other "scores," etc.}

\item{...}{Arguments passed on to other plot/points functions}

\item{GenomewideLine}{Where to draw a "genome-wide sigificant" line. Default
-log10(5e-8). Set to FALSE to disable.}

\item{HighLight}{A character vector of SNPs in your dataset to highlight.
These SNPs should all be in your dataset.}
}
\value{
A manhattan plot file of pdf or png.
}
\description{
Creates a manhattan plot from PLINK assoc output (or any data frame with
chromosome, position, and p-value).
}
\examples{
manhattan(gwasResults)

}
\keyword{manhattan}
\keyword{visualization}
