% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASC.R
\name{MASC}
\alias{MASC}
\title{Matching and syntehtic control}
\usage{
MASC(Y, X, space, K, phi)
}
\arguments{
\item{Y}{The outcome (post-treatment) variables for the control group}

\item{X}{The features on which to match, usually pre-treatment outcomes}

\item{space}{Space on which to predict, typically pre-treatment for treated group}

\item{K}{The number of matches to use}

\item{phi}{The mixture tuning parameter}
}
\description{
Estimates mixture of matching and synthetic control estimates
}
