% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{matching}
\alias{matching}
\title{Matching estimation for multivariable independent and dependent variables}
\usage{
matching(Y, X, space, K = 1)
}
\arguments{
\item{Y}{The outcome variables (can be multi-valued as in panel data)}

\item{X}{The explanatory variables}

\item{space}{The space on whic to predict, for example treated group's X}

\item{K}{Number of matches to use}
}
\description{
Requires Gurobi optimizer
}
