% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifyFilesWithSameName.R
\name{ModifyAndBackup}
\alias{ModifyAndBackup}
\title{Modify a specified file with fixed file name}
\usage{
ModifyAndBackup(
  Dir,
  FileName,
  KeyWordOldLocation,
  KeyWordOld,
  KeyWordNew,
  Replace = TRUE,
  SaveOld = TRUE
)
}
\arguments{
\item{Dir}{The directory where you target files exist.}

\item{FileName}{complete file name.}

\item{KeyWordOldLocation}{Key word used to find target lines.}

\item{KeyWordOld}{The old word needs to be replaced.}

\item{KeyWordNew}{The new word.}

\item{Replace}{Whether to replace the old file. Default is FALSE.}

\item{SaveOld}{Wheter to save the old file, this will suffix date to the file name.}
}
\value{

}
\description{
Modify a specified file with fixed file name
}
