% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifyFilesWithKeyWrodsInFileName.R
\name{ModifyAndBackupFile}
\alias{ModifyAndBackupFile}
\title{Modify And Backup File}
\usage{
ModifyAndBackupFile(
  poss.files,
  KeyWordOldLocation,
  ignore.case,
  KeyWordOld,
  KeyWordNew,
  Replace = TRUE,
  SaveOld = TRUE
)
}
\arguments{
\item{poss.files}{The target file.}

\item{KeyWordOldLocation}{key words used to find target line, could be a character vector.}

\item{ignore.case}{whether to ignore case of line key words for searching files. default is FALSE.}

\item{Replace}{Whether to replace the old file. Default is FALSE.}

\item{SaveOld}{Wheter to save the old file, this will suffix date to the file name.}

\item{keyWordOld}{The old word needs to be replaced.}

\item{keyWordNew}{The new word.}
}
\value{

}
\description{
Modify And Backup File
}
