% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Classes.R
\docType{class}
\name{MyDataSet-class}
\alias{MyDataSet-class}
\title{The main experiment-level class for MyDataSet data}
\description{
Contains all currently-supported component data classes:
\code{\link{Profile_table-class}},
\code{\link{Sample_data-class}},
\code{\link{Feature_table-class}},
\code{MyDataSet-class} object the only data argument required for analysis and plotting
functions (although there are many options and parameter arguments available
to you).
}
\details{
In the case of missing component data, the slots are set to \code{NULL}. As
soon as a \code{MyDataSet-class} object is to be updated with new component
data (previously missing/\code{NULL} or not), the indices of all components
are re-checked for compatibility and trimmed if necessary. This is to ensure
by design that components describe the same features/samples, and also that these
trimming/validity checks do not need to be repeated in downstream analyses.

slots:
\describe{
   \item{Profile_table}{a single object of class Profile_table}
   \item{Sample_data}{ a single object of class Sample_data}
   \item{Feature_table}{ a single object of class Feature_table}
}
}
