% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRTools.R
\name{myRTools.fread}
\alias{myRTools.fread}
\title{myRTools Fast and friendly file finagler}
\usage{
myRTools.fread(...)
}
\arguments{
\item{...}{XXX}
}
\value{
A data.table by default. A data.frame when argument data.table=FALSE;
e.g. options(datatable.fread.datatable=FALSE).
}
\description{
myRTools.fread is an extension of \code{\link[data.table]{fread}} for
convenient R programming. In the first line, it stores classes of \code{x}
followed by \code{x} itselfs.
}
\details{
myRTools.fread is an extension of \code{\link[data.table]{fread}} for
convenient R programming. In the first line, it stores classes of \code{x}
followed by \code{x} itselfs.
\code{\link[myRTools]{myRTools.fwrite}}. Afterwards it load the rest of the
file and converts all columns according to the class definition.
}
\examples{
dateToday1 <- today()
timeNow1 <- now()
dt1 <- data.table(int = 1:5,
                 num = seq(0.2, 1, by = 0.2),
                 char = c("a", "b", "c", "d", "f"),
                 date = dateToday1 + days(0:4),
                 time = timeNow1 + days(0:4))
myRTools.fwrite(dt1, file = "test.csv")

dateToday2 <- today() - years(1)
timeNow2 <- now() - years(1)
dt2 <- data.table(int = 6:10,
                  num = seq(1.2, 2, by = 0.2),
                  char = c("v", "w", "x", "y", "z"),
                  date = dateToday2 + days(0:4),
                  time = timeNow2 + days(0:4))
myRTools.fwrite(dt2, file = "test.csv", append = TRUE)

dt3 <- myRTools.fread(file="test.csv")
print(dt3)
}
\seealso{
\code{\link[myRTools]{myRTools.fwrite}}, \code{\link[data.table]{fread}}, \code{\link[fasttime]{fastPOSIXct}}, \code{\link[lubridate]{as_date}}
}
\author{
Christian Frei
}
