% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRTools.R
\name{myRTools.fwrite}
\alias{myRTools.fwrite}
\title{myRTools Fast CSV writer}
\usage{
myRTools.fwrite(x, file, append = FALSE, ...)
}
\arguments{
\item{x}{Any list of same length vectors; e.g. data.frame and data.table.}

\item{file}{Output file name. "" indicates output to the console.}

\item{append}{If TRUE, the file is opened in append mode and column names (header row) are not written.}

\item{...}{XXX}
}
\description{
myRTools.fwrite is an extension of \code{\link[data.table]{fwrite}} for
convenient R programming. In the first line, it stores classes of \code{x}
followed by \code{x} itselfs.
}
\details{
myRTools.fwrite is an extension of \code{\link[data.table]{fwrite}}
for convenient R programming. In the first line, it stores classes of
\code{x} followed by \code{x} itselfs.
}
\examples{
dateToday1 <- today()
timeNow1 <- now()
dt1 <- data.table(int = 1:5,
                 num = seq(0.2, 1, by = 0.2),
                 char = c("a", "b", "c", "d", "f"),
                 date = dateToday1 + days(0:4),
                 time = timeNow1 + days(0:4))
myRTools.fwrite(dt1, file = "test.csv")

dateToday2 <- today() - years(1)
timeNow2 <- now() - years(1)
dt2 <- data.table(int = 6:10,
                  num = seq(1.2, 2, by = 0.2),
                  char = c("v", "w", "x", "y", "z"),
                  date = dateToday2 + days(0:4),
                  time = timeNow2 + days(0:4))
myRTools.fwrite(dt2, file = "test.csv", append = TRUE)

}
\seealso{
\code{\link[myRTools]{myRTools.fread}} \code{\link[data.table]{fwrite}}
}
\author{
Christian Frei
}
