% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.neighbors.1d.R
\name{nearest.neighbors.1d}
\alias{nearest.neighbors.1d}
\title{Nearest neighbors in one dimension}
\usage{
nearest.neighbors.1d(Y, X, k = 1, space)
}
\arguments{
\item{Y}{A vector of outcomes}

\item{X}{A vector of predictors}

\item{k}{The number of neighbors to use}

\item{space}{The values at which to predict}
}
\description{
This performs nearest neighbors estimation in one dimension. For higher
dimensional analysis, this can be the propensity score.
}
