% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_DataAccess.R
\docType{methods}
\name{nfeature}
\alias{nfeature}
\alias{nfeature,ANY-method}
\alias{nfeature,MyDataSet-method}
\alias{nfeature,Profile_table-method}
\alias{nfeature,Feature_table-method}
\title{Get the number of feature.}
\usage{
nfeature(mydataset)

\S4method{nfeature}{ANY}(mydataset)

\S4method{nfeature}{MyDataSet}(mydataset)

\S4method{nfeature}{Profile_table}(mydataset)

\S4method{nfeature}{Feature_table}(mydataset)
}
\arguments{
\item{mydataset}{\code{\link{MyDataSet-class}}, \code{\link{Profile_table-class}},
\code{\link{Feature_table-class}}}
}
\value{
An integer indicating the number of feature.
}
\description{
Get the number of feature.
}
\examples{

}
\seealso{
feature_names
}
