% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_normalization.R
\name{norm_tss}
\alias{norm_tss}
\alias{norm_tmm}
\alias{norm_rle}
\alias{norm_clr}
\alias{norm_zscore}
\alias{norm_median}
\alias{norm_mad}
\alias{norm_robust}
\alias{norm_unit}
\alias{norm_minmax}
\title{Total-Sum Scaling (TSS) method}
\usage{
norm_tss(object)

norm_tmm(
  object,
  ref_column = NULL,
  logratio_trim = 0.3,
  sum_trim = 0.05,
  do_weighting = TRUE,
  Acutoff = -1e+10
)

norm_rle(
  object,
  locfunc = stats::median,
  type = c("poscounts", "ratio"),
  geo_means = NULL,
  control_genes = NULL
)

norm_clr(object)

norm_zscore(object)

norm_median(object)

norm_mad(object)

norm_robust(object)

norm_unit(object)

norm_minmax(object)
}
\arguments{
\item{object, }{Object; a [`matrix`](Row->Features; Column->Samples).}

\item{ref_column}{column to use as reference}

\item{logratio_trim}{amount of trim to use on log-ratios}

\item{sum_trim}{amount of trim to use on the combined absolute levels
("A" values)}

\item{do_weighting}{whether to compute the weights or not}

\item{Acutoff}{cutoff on "A" values to use before trimming}

\item{locfunc}{a function to compute a location for a sample. By default,
the median is used.}

\item{type}{method for estimation: either "ratio"or "poscounts" (recommend).}

\item{geo_means}{default `NULL`, which means the geometric means of the
counts are used. A vector of geometric means from another count matrix can
be provided for a "frozen" size factor calculation.}

\item{control_genes}{default `NULL`, which means all taxa are used for size
factor estimation, numeric or logical index vector specifying the taxa
used for size factor estimation (e.g. core taxa).}
}
\description{
TSS simply transforms the feature table into relative abundance by dividing
the number of total reads of each sample.

TMM calculates the normalization factor using a robust statistics based on
the assumption that most features are not differential and should, in
average, be equal between the samples. The TMM scaling factor is calculated
as the weighted mean of log-ratios between each pair of samples, after
excluding the highest count OTUs and OTUs with the largest log-fold change.

RLE assumes most features are not differential and uses the relative
abundances to calculate the normalization factor.

In CLR, the log-ratios are computed relative to the geometric mean of all
features.
}
\seealso{
[edgeR::calcNormFactors()]

[DESeq2::estimateSizeFactorsForMatrix()]
}
