% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{NormalizeFun}
\alias{NormalizeFun}
\title{Normalization}
\usage{
NormalizeFun(Vector, type="Zscore")
}
\arguments{
\item{Vectors, }{Numeric; (Required) a vector contains numeric.}

\item{type, }{Character; "MAD": Median Absolute Deviation; "Robust"; "Unit"; "Zscore"; "Median"; "Min_Max" scale normalization(default: Zscore).}
}
\value{
a normalized vector
}
\description{
Multiple variables may have different, leading to unequal contribution on the data analysis. Normalization could figure out this drawback when we do some predictive analysis.
}
\details{
12/1/2021 Guangzhou China
}
\examples{

V <- c(1:20, 30:40)
M <- matrix(rnorm(24, mean = 10, sd=10), nrow=4, ncol=6, byrow=TRUE, dimnames = list(paste0("R", 1:4), paste0("C", 1:6)))
V_norm <- NormalizeFun(V, type="Zscore")
M_norm <- t(apply(M, 1, NormalizeFun, "Zscore"))
}
\author{
Hua Zou
}
