% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_DataAccess.R
\docType{methods}
\name{nsamples}
\alias{nsamples}
\alias{nsamples,ANY-method}
\alias{nsamples,MyDataSet-method}
\alias{nsamples,Profile_table-method}
\alias{nsamples,Sample_data-method}
\title{Get the number of samples.}
\usage{
nsamples(mydataset)

\S4method{nsamples}{ANY}(mydataset)

\S4method{nsamples}{MyDataSet}(mydataset)

\S4method{nsamples}{Profile_table}(mydataset)

\S4method{nsamples}{Sample_data}(mydataset)
}
\arguments{
\item{mydataset}{A \code{\link{MyDataSet-class}}, \code{\link{Sample_data}},
or \code{\link{Profile_table-class}}.}
}
\value{
An integer indicating the total number of samples.
}
\description{
Get the number of samples.
}
\examples{

}
\seealso{
\code{\link{feature_names}}, \code{\link{sample_names}},
 \code{\link{nfeature}}
}
