% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{om_box}
\alias{om_box}
\title{boxplot}
\usage{
om_box(plot_data, samples, outdir)
}
\arguments{
\item{plot_data}{input data which will be ploted}

\item{samples}{a data frame about samples infomation}

\item{outdir}{a character path to output}
}
\description{
onmath quantification analysis boxplot
}
