% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant.R
\name{om_heatmap}
\alias{om_heatmap}
\title{heatmap}
\usage{
om_heatmap(plot_data, samples, outdir)
}
\arguments{
\item{plot_data}{a matrix which will be ploted}

\item{samples}{a data frame about samples infomation}

\item{outdir}{a character path to output}
}
\description{
onmath quantification analysis cluster heatmap
}
