% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_Ordination.R
\name{Ord_PCA}
\alias{Ord_PCA}
\title{Principal Component Analysis}
\usage{
Ord_PCA(dataset=ExpressionSet, Normalization=TRUE, Group_info="Group")
}
\arguments{
\item{Normalization, }{Logical; Normalizing feature(default: TRUE).}

\item{Group_info, }{Character; the group for plot(default: "Group").}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list object:
  PCA score
  Result of PERMANOVA
}
\description{
The common multimension reduction method is Principal Component Analysis (PCA). It's suitable for non-zero sparse matrix such as metabolites' profile.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

data(ExprSet_species)

PCA_res <- Ord_PCA(dataset=ExprSet_species, Normalization=TRUE, Group_info="Group")
PCA_res$PCA

}
\author{
Hua Zou
}
