% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_Ordination.R
\name{Ord_PCoA}
\alias{Ord_PCoA}
\title{Principal Coordinate Analysis}
\usage{
Ord_PCoA(dataset=ExpressionSet, Method="bray", Group_info="Group")
}
\arguments{
\item{Method, }{Character; method to calculate the distance(default: bray).}

\item{Group_info, }{Character; the group for plot(default: "Group").}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list object:
  PCoA score
  Result of PERMANOVA
}
\description{
Principal Coordinate Analysis use the distance among samples, which calculates through the multiple variables.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

data(ExprSet_species)

PCoA_res <- Ord_PCoA(dataset=ExprSet_species, Method="bray", Group_info="Group")
PCoA_res$PCoA

}
\author{
Hua Zou
}
