% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_Ordination.R
\name{Ord_Plot}
\alias{Ord_Plot}
\title{Plot Ordination result with scatterplot}
\usage{
Ord_Plot(Score=score, Axis_name=explains, Pvalue=label)
}
\arguments{
\item{Score, }{Data.frame; (Required) Result of Ordination.}

\item{Axis_name, }{Character; the labels of two axis.}

\item{Pvalue, }{Character; (Required) PERMANOVA result.}
}
\value{
a scatterplot
}
\description{
Show the result of Ordination using scatterplot.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

data(ExprSet_species)

PCA_res <- Ord_PCA(dataset=ExprSet_species, Normalization=TRUE, Group_info="Group")
Ord_Plot(Score=PCA_res$PCA, Axis_name=PCA_res$epn, Pvalue=PCA_res$PER)

res_tsne <- Ord_tsne(dataset=ExprSet_species, Perplexity=20, Group_info="Group")
Ord_Plot(Score=res_tsne$tsne, Axis_name=res_tsne$epn, Pvalue=res_tsne$PER)


}
\author{
Hua Zou
}
