% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06.application.R
\name{pca_scatterplot}
\alias{pca_scatterplot}
\title{PCA_scatterplot}
\usage{
pca_scatterplot(pca_site)
}
\description{
the scatterplot for PCA
 geom_enterotype function was from https://stackoverflow.com/questions/42575769/how-to-modify-the-backgroup-color-of-label-in-the-multiple-ggproto-using-ggplot2
}
\details{
08/01/2020  ShenZhen China
}
\examples{
 pca_plot <- pca_scatterplot(pca_site)

}
\author{
Hua Zou
}
