% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07.association.R
\name{PERMANOVA}
\alias{PERMANOVA}
\title{PERMANOVA}
\usage{
PERMANOVA(phen, prof)
}
\description{
PERMANOVA
PERMANOVA the over gut microbial species and each phenotype
}
\details{
08/01/2020  ShenZhen China
}
\examples{
 per_res <- PERMANOVA(phen, prof)

}
\author{
Hua Zou
}
