% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit.R
\name{probit}
\alias{probit}
\title{Probit estimation}
\usage{
probit(Y, X, const = TRUE)
}
\arguments{
\item{Y}{The outcome variables}

\item{X}{The explanatory variables

NOTE: Add variance later. documentation is funny}
}
\description{
Estimates probit parameters and fitted values
}
