% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Classes.R
\docType{class}
\name{Profile_table-class}
\alias{Profile_table-class}
\title{The S4 class for storing Profile_table information.}
\description{
Because orientation of these tables can vary by method, the orientation is
defined explicitly in the \code{feature_are_rows} slot (a logical).
The \code{Profile_table} class inherits the \code{\link{matrix}} class to store
expression values.
Various standard subset and assignment nomenclature has been extended to apply
to the \code{Profile_table} class, including square-bracket, \code{\link{t}}, etc.
}
\details{
\describe{
   \item{feature_are_rows}{
	A single logical specifying the orientation of the expression table.
   }

   \item{.Data}{This slot is inherited from the \code{\link{matrix}} class.}
 }
}
