% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_DataAccess.R
\docType{methods}
\name{Profile_table}
\alias{Profile_table}
\alias{Profile_table,MyDataSet-method}
\alias{Profile_table,Profile_table-method}
\alias{Profile_table,matrix-method}
\alias{Profile_table,data.frame-method}
\alias{Profile_table,ANY-method}
\title{Build or access the Profile_table}
\usage{
Profile_table(object, feature_are_rows, errorIfNULL=TRUE)

\S4method{Profile_table}{MyDataSet}(object, errorIfNULL = TRUE)

\S4method{Profile_table}{Profile_table}(object, errorIfNULL = TRUE)

\S4method{Profile_table}{matrix}(object, feature_are_rows)

\S4method{Profile_table}{data.frame}(object, feature_are_rows)

\S4method{Profile_table}{ANY}(object, errorIfNULL = TRUE)
}
\arguments{
\item{object}{(Required). An integer matrix, \code{\link{Profile_table-class}},
or \code{\link{MyDataSet-class}}.}

\item{feature_are_rows}{(Conditionally optional). Logical; of length 1. Ignored
unless \code{object} is a matrix, in which case it is is required.}

\item{errorIfNULL}{(Optional). Logical. Should the accessor stop with
an error if the slot is empty (\code{NULL})? Default \code{TRUE}. Ignored
if \code{object} argument is a matrix (constructor invoked instead).}
}
\value{
An \code{\link{Profile_table-class}} object.
}
\description{
This is the suggested method for both constructing and accessing
Operational Feature profile (\code{\link{Profile_table-class}}) objects.
When the first
argument is a matrix, Profile_table() will attempt to create and return an
Profile_table-class object,
which further depends on whether or not \code{feature_are_rows} is provided as an
additional argument.
Alternatively, if the first argument is an experiment-level (\code{\link{MyDataSet-class}})
object, then the corresponding \code{Profile_table} is returned.
}
\examples{

}
