% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_MyDataSet.R
\name{prune_feature}
\alias{prune_feature}
\alias{prune_feature,NULL,ANY-method}
\alias{prune_feature,logical,ANY-method}
\alias{prune_feature,character,Profile_table-method}
\alias{prune_feature,character,Sample_data-method}
\alias{prune_feature,character,sample_data-method}
\alias{prune_feature,character,MyDataSet-method}
\alias{prune_feature,character,Feature_table-method}
\title{Prune unwanted Features from a MyDataSet object.}
\usage{
prune_feature(feature, x)

\S4method{prune_feature}{`NULL`,ANY}(feature, x)

\S4method{prune_feature}{logical,ANY}(feature, x)

\S4method{prune_feature}{character,Profile_table}(feature, x)

\S4method{prune_feature}{character,Sample_data}(feature, x)

\S4method{prune_feature}{character,MyDataSet}(feature, x)

\S4method{prune_feature}{character,Feature_table}(feature, x)
}
\arguments{
\item{feature}{(Required). A character vector of the feature in object x that you want to
keep -- OR alternatively -- a logical vector where the kept feature are TRUE, and length
is equal to the number of feature in object x. If \code{feature} is a named
logical, the feature retained are based on those names. Make sure they are
compatible with the \code{feature_names} of the object you are modifying (\code{x}).}

\item{x}{(Required). A MyDataSet object, including MyDataSet classes that represent feature.
If the function \code{\link{feature_names}} returns a non-\code{NULL} value, then your object
can be pruned by this function.}
}
\value{
The class of the object returned by \code{prune_feature} matches
the class of the argument, \code{x}.
}
\description{
An S4 Generic method for removing (pruning) unwanted Features from MyDataSet
objectsas well as native MyDataSet package objects.
This is particularly useful for pruning a MyDataSet object that has
more than one component that describes Features.
Credit: the \code{phylo}-class version is adapted from
\href{http://cran.at.r-project.org/web/packages/picante/index.html}{prune.sample}.
}
\examples{

}
\seealso{
\code{\link{prune_samples}}

 \href{http://cran.at.r-project.org/web/packages/picante/index.html}{prune.sample}
}
