% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_MyDataSet.R
\docType{methods}
\name{prune_samples}
\alias{prune_samples}
\alias{prune_samples,character,Profile_table-method}
\alias{prune_samples,character,Sample_data-method}
\alias{prune_samples,character,MyDataSet-method}
\alias{prune_samples,logical,ANY-method}
\title{Define a subset of samples to keep in a MyDataSet object.}
\usage{
prune_samples(samples, x)

\S4method{prune_samples}{character,Profile_table}(samples, x)

\S4method{prune_samples}{character,Sample_data}(samples, x)

\S4method{prune_samples}{character,MyDataSet}(samples, x)

\S4method{prune_samples}{logical,ANY}(samples, x)
}
\arguments{
\item{samples}{(Required). A character vector of the samples in object x that you want to
keep -- OR alternatively -- a logical vector where the kept samples are TRUE, and length
is equal to the number of samples in object x. If \code{samples} is a named
logical, the samples retained is based on those names. Make sure they are
compatible with the \code{sample_names} of the object you are modifying (\code{x}).}

\item{x}{A MyDataSet object.}
}
\value{
The class of the object returned by \code{prune_samples} matches
the class of the MyDataSet object, \code{x}.
}
\description{
An S4 Generic method for pruning/filtering unwanted samples
by defining those you want to keep.
}
\examples{

}
