% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_DataAccess.R
\name{reconcile_categories}
\alias{reconcile_categories}
\title{Cleans absent levels in Sample_data/data.frame.}
\usage{
reconcile_categories(DFSM)
}
\arguments{
\item{DFSM}{(Required). A \code{data.frame} or \code{Sample_data} object that needs to be cleaned.}
}
\value{
A single \code{data.frame} object. Even if the input argument is a \code{Sample_data},
 the return is a \code{data.frame}. Because this is intended to be used internally by
 the builder method, it cannot also call the builder function to re-build
 the cleaned \code{Sample_data}.
}
\description{
This is used internally by the builder method, \code{\link{Sample_data}}, to
ensure that the factors describing categorical variables in a data.frame or
Sample_data object are free of extra levels that can plague downstream plots
analysis.
}
\examples{

}
\keyword{internal}
