% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_DESeq2.R
\name{run_DESeq2}
\alias{run_DESeq2}
\title{Differential Expression Analysis by DEseq2 Package}
\usage{
run_DESeq2(dataset=ExpressionSet,
                  trim="none",
                  Group_info="Group",
                  Group_name=c("HC", "AA"),
                  Pvalue=0.05, Log2FC=1)
}
\arguments{
\item{trim, }{Character; filter to apply.(default: trim="none").}

\item{Group_info, }{Character; design factor(default: "Group").}

\item{Group_name, }{Character; (Required) the group for comparison.}

\item{Pvalue, }{Numeric; significant level(default: 0.05).}

\item{Log2FC, }{Numeric; log2FoldChange(default: 1).}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list object:
  DESeq object
  DESeq results
  significant difference with enriched directors
}
\description{
DEseq2 requires the count data (a matrix of integer values) to input. The normalization method is to use the standard factor vector per feature.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSetRawCount)

DESeq2_res <- run_DESeq2(dataset=ExprSetRawCount, Group_info="Group", Group_name=c("HC", "AA"), Pvalue=0.05, Log2FC=1)
DESeq2_res$res
}

}
\author{
Hua Zou
}
