% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_filter.R
\name{run_filter}
\alias{run_filter}
\title{Filter samples or features in `assayData` by Occurrences}
\usage{
run_filter(object, object=0.2, filterType="Both")
}
\arguments{
\item{object, }{Object; a [`matrix`] or [`assayData-class`] or [`ExpressionSet-class`].}

\item{cutoff, }{Numeric; the threshold for filtering (default: Cutoff=0.2).}

\item{filterType, }{Character; the type of filtering data ("identity", "both", "feature", "sample", "Group").}
}
\value{
A filtered `object` with the cutoff
}
\description{
Filter samples or features in `assayData` by Occurrences,
which means the samples or features will be discard if they could not pass the cutoff.
}
\details{
12/6/2021 Guangzhou China
}
\examples{
\donttest{
   data("ExprSet_species")
   run_filter(ExprSet_species, object=0.2, filterType="both")
}

}
\author{
Hua Zou
}
