% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_imputation.R
\name{run_impute}
\alias{run_impute}
\title{impute the expression profile in `assayData` by features}
\usage{
run_impute(object, normalize=c("none", "deletion", "GlobalStandard", "KNN"))
}
\arguments{
\item{object, }{Object; a [`matrix`] or [`assayData-class`] or [`ExpressionSet-class`].}

\item{normalize, }{Character; normalization to apply, the options inclulde:
* "none": return the original data without any transformation.
* "deletion": the deletion method is used when the probability of missing variable is same for all observations.
* "GlobalStandard": (Mean/ Mode/ Median Imputation), consists of replacing the missing data for a given attribute by the mean or median (quantitative attribute) or mode (qualitative attribute) of all known values of that variable.
* "KNN": the missing values of an attribute are imputed using the given number of attributes that are most similar to the attribute whose values are missing.}
}
\value{
A object matches the class of argument `object` with the imputated profile.
}
\description{
The missing values in expression profile maybe affect the statistical results, imputing the NAs or Zero values should taken into account.
}
\details{
12/13/2021 Guangzhou China
}
\examples{

}
\author{
Hua Zou
}
