% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_limma_voom.R
\name{run_Limma}
\alias{run_Limma}
\title{Differential Expression Analysis by limma Package}
\usage{
run_Limma(dataset=ExpressionSet,
                 trim="none",
                 Group_info="Group",
                 Group_name=c("HC", "AA"),
                 Pvalue=0.05,
                 Log2FC=1)
}
\arguments{
\item{trim, }{Character; filter to apply.(default: trim="none").}

\item{Group_info, }{Character; design factor(default: "Group").}

\item{Group_name, }{Character; (Required) the group for comparison.}

\item{Pvalue, }{Numeric; significant level(default: 0.05).}

\item{Log2FC, }{Numeric; log2FoldChange(default: 1).}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list object:
  limma results
  significant difference with enriched directors
}
\description{
limma requires the count data (a matrix of integer values) to input. The *Voom* would transform the count data for further data analysis.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSetRawCount)

Limma_res <- run_Limma(dataset=ExprSetRawCount, Group_info="Group", Group_name=c("HC", "AA"), Pvalue=0.05, Log2FC=1)
Limma_res$res
}

}
\author{
Hua Zou
}
