% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_Ordination.R
\name{run_MDS}
\alias{run_MDS}
\title{Dimensionality Reduction Analysis: Multidimensional Scaling(MDS)}
\usage{
run_MDS(dataset=ExpressionSet, Method="Classic", Group_info="Group")
}
\arguments{
\item{Method, }{Character; method for MDS (default: Classic).}

\item{Group_info, }{Character; the group for plot(default: "Group").}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list object:
  MDS score
  Result of PERMANOVA
}
\description{
MDS use the distance among samples, which calculates through the multiple variables.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSetRawRB)

MDS_res <- run_MDS(dataset=ExprSetRawRB, Method="Classic", Group_info="Group")
MDS_res$MDS
}

}
\author{
Hua Zou
}
