% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{run_OddRatio}
\alias{run_OddRatio}
\title{Obtain the 95% CI Odd Ratio by glm model}
\usage{
run_OddRatio(datx, daty, GroupName)
}
\arguments{
\item{datx, }{Data.frame; (Required) Metadata with SampleID and Group information.}

\item{daty, }{Data.frame; (Required) Profile with SampleID, ordered by datx (Row->Features; Column->SampleID).}

\item{GroupName, }{Character; (Required) the contrast group.}
}
\value{
95% CI Odd Ratio with featureID
}
\description{
Obtain the 95% CI Odd Ratio by glm model
}
\details{
12/13/2021 Guangzhou China
}
\examples{

\donttest{

}

}
\author{
Hua Zou
}
