% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_Ordination.R
\name{run_PCA}
\alias{run_PCA}
\title{Dimensionality Reduction Analysis: Principal Component Analysis(PCA)}
\usage{
run_PCA(dataset=ExpressionSet, Normalize="Zscore", Group_info="Group")
}
\arguments{
\item{Normalize, }{Character; Normalizing feature(default: normalize="none").}

\item{Group_info, }{Character; the group for plot(default: "Group").}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list object:
  PCA score
  Result of PERMANOVA
}
\description{
The common Dimensionality reduction method is PCA. It's suitable for non-zero sparse matrix such as metabolites' profile.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{
data("ExprSetRawRB")

PCA_res <- run_PCA(dataset=ExprSetRawRB, Normalize="Zscore", Group_info="Group")
PCA_res$PCA
}

}
\author{
Hua Zou
}
