% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_T_Test.R
\name{run_Ttest_Paired}
\alias{run_Ttest_Paired}
\title{Differential Expression Analysis by Paired T test}
\usage{
run_Ttest_Paired(dataset=ExpressionSet,
                  trim="none",
                  transform="none",
                  normalize="none",
                  Group_info="Group",
                  Group_name=c("HC", "AA"),
                  Pair_ID="PID",
                  Pvalue=0.05,
                  Log2FC=1)
}
\arguments{
\item{trim, }{Character; filter to apply.(default: trim="none").}

\item{transform, }{Character; transformation to apply.(default: tranform="none").}

\item{normalize, }{Character; normalization to apply.(default: normalize="none").}

\item{Group_info, }{Character; design factor(default: "Group").}

\item{Group_name, }{Character; (Required) the group for comparison.}

\item{Pair_ID, }{Character; (Required) the paired ID.}

\item{Pvalue, }{Numeric; significant level(default: 0.05).}

\item{Log2FC, }{Numeric; log2FoldChange(default: 1).}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
significant difference with enriched directors:
* Features
* Block
* Enrichment
* AdjustPvalue
* Pvalue
* Log2FC
* Statistic
* Median (All/Groups)
* Odds Ratio (95% CI)
}
\description{
Paired T test is parameter test method, and also use for the data with normal distribution(Gauss Distribution).
}
\details{
12/31/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSetRawCount)

t_res <- run_Ttest_Paired(dataset=ExprSetRawCount, Group_info="Group", Group_name=c("HC", "AA"), Pair_ID="PID", Pvalue=0.05, Log2FC=1)
t_res$res
}

}
\author{
Hua Zou
}
