% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_Ordination.R
\name{run_umap}
\alias{run_umap}
\title{Dimensionality Reduction Analysis: Uniform Manifold Approximation and Projection (UMAP)}
\usage{
run_umap(dataset=ExpressionSet, Group_info="Group")
}
\arguments{
\item{Group_info, }{Character; the group for plot(default: "Group").}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
a list object:
  umap score
  Result of PERMANOVA
}
\description{
UMAP: a non-linear dimensionality reduction algorithm.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSetRawRB)

umap_res <- run_umap(dataset=ExprSetRawRB, Group_info="Group")
umap_res$umap
}

}
\author{
Hua Zou
}
