% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_WilcoxRankSumTest.R
\name{run_Wilcox}
\alias{run_Wilcox}
\title{Differential Expression Analysis by Wilcox-rank-sum test}
\usage{
run_Wilcox(dataset=ExpressionSet,
                  trim="none",
                  transform="none",
                  normalize="none",
                  Group_info="Group",
                  Group_name=c("HC", "AA"),
                  Pvalue=0.05,
                  Log2FC=1)
}
\arguments{
\item{trim, }{Character; filter to apply.(default: trim="none").}

\item{transform, }{Character; transformation to apply.(default: tranform="none").}

\item{normalize, }{Character; normalization to apply.(default: normalize="none").}

\item{Group_info, }{Character; design factor(default: "Group").}

\item{Group_name, }{Character; (Required) the group for comparison.}

\item{Pvalue, }{Numeric; significant level(default: 0.05).}

\item{Log2FC, }{Numeric; log2FoldChange(default: 1).}

\item{Expression, }{ExpressionSet; (Required) ExpressionSet object.}
}
\value{
significant difference with enriched directors:
* Features
* Block
* Enrichment
* AdjustPvalue
* Pvalue
* Log2FC
* Statistic
* Median (All/Groups)
* Odds Ratio (95% CI)
}
\description{
Wilcox-rank-sum test is non-parameter test method, and also use for the data with non-normal distribution(Gauss Distribution).
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSetRawCount)

t_res <- run_Wilcox(dataset=ExprSetRawCount, Group_info="Group", Group_name=c("HC", "AA"), Pvalue=0.05, Log2FC=1)
t_res$res
}

}
\author{
Hua Zou
}
