% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_DataAccess.R
\docType{methods}
\name{Sample_data}
\alias{Sample_data}
\alias{Sample_data,ANY-method}
\alias{Sample_data,data.frame-method}
\title{Build or access Sample_data.}
\usage{
Sample_data(object, errorIfNULL=TRUE)

\S4method{Sample_data}{ANY}(object, errorIfNULL = TRUE)

\S4method{Sample_data}{data.frame}(object)
}
\arguments{
\item{object}{(Required). A \code{\link{data.frame-class}},
or a \code{\link{MyDataSet-class}} object.}

\item{errorIfNULL}{(Optional). Logical. Should the accessor stop with
an error if the slot is empty (\code{NULL})? Default \code{TRUE}.}
}
\value{
A \code{\link{Sample_data-class}} object
representing the sample variates of an experiment.
}
\description{
This is the suggested method for both constructing and accessing a table
of sample-level variables (\code{\link{Sample_data-class}}),
which in the \code{\link{MyDataSet-package}} is represented as a special
extension of the \code{\link{data.frame-class}}.
When the
argument is a \code{\link{data.frame}}, \code{Sample_data} will create
a Sample_data-class object.
In this case, the rows should be named to match the
\code{\link{sample_names}} of the other objects to which it will ultimately be paired.
Alternatively, if the first argument is an experiment-level (\code{\link{MyDataSet-class}})
object, then the corresponding \code{Sample_data} is returned.
Like other accessors (see See Also, below), the default behavior of this method
is to stop with an
error if \code{object} is a \code{MyDataSet-class} but does not
contain a \code{Sample_data}.
}
\examples{

}
