% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_DataAccess.R
\docType{methods}
\name{sample_names}
\alias{sample_names}
\alias{sample_names,ANY-method}
\alias{sample_names,MyDataSet-method}
\alias{sample_names,Sample_data-method}
\alias{sample_names,Profile_table-method}
\title{Get sample names.}
\usage{
sample_names(mydataset)

\S4method{sample_names}{ANY}(mydataset)

\S4method{sample_names}{MyDataSet}(mydataset)

\S4method{sample_names}{Sample_data}(mydataset)

\S4method{sample_names}{Profile_table}(mydataset)
}
\arguments{
\item{mydataset}{(Required). A \code{\link{MyDataSet-class}}, \code{\link{Sample_data}},
or \code{\link{Profile_table-class}}.}
}
\value{
A character vector. The names of the samples in \code{mydataset}.
}
\description{
Get sample names.
}
\examples{

}
