% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_packages.R
\name{save_gtsummary}
\alias{save_gtsummary}
\title{Save a gtsummary Table to File}
\usage{
save_gtsummary(tbl, filename)
}
\arguments{
\item{tbl}{A gtsummary table object}

\item{filename}{Output filename (.html, .pdf, or .docx)}
}
\value{
Invisible NULL
}
\description{
Saves a gtsummary table to HTML, PDF, or Word format.
}
\examples{
\dontrun{
tbl <- make_desc_or_table(dat_filter, refusal, vars)
save_gtsummary(tbl, "Table1.html")
save_gtsummary(tbl, "Table1.docx")
}
}
