% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelloWorld.R
\name{sayHi}
\alias{sayHi}
\title{say hi to you input name}
\usage{
sayHi(name)
}
\arguments{
\item{name}{input a character type}
}
\value{
character hi,'your name'
}
\description{
say hi to you input name
}
\examples{
sayHi("chencheng")
}
\keyword{name}
