% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.visualization.R
\name{scatterplot1}
\alias{scatterplot1}
\title{scatterplot1}
\usage{
scatterplot1(dat = DATA2)
}
\arguments{
\item{DATA}{Table.}
}
\value{
scatterplot
}
\description{
scatterplot1
}
\examples{
scatterplot1(DATA2)

}
