% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_DataAccess.R
\name{show,Profile_table-method}
\alias{show,Profile_table-method}
\alias{show,Sample_data-method}
\alias{show,Feature_Table-method}
\alias{show,MyDataSet-method}
\title{method extensions to show for MyDataSet objects.}
\usage{
\S4method{show}{Profile_table}(object)

\S4method{show}{Sample_data}(object)

\S4method{show}{Feature_Table}(object)

\S4method{show}{MyDataSet}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
See the general documentation of \code{\link[methods]{show}} method for
expected behavior.
}
\examples{

}
\seealso{
\code{\link[methods]{show}}
}
