% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sieve.R
\name{sieve}
\alias{sieve}
\title{Sieve}
\usage{
sieve(Y, X, order = 1, space)
}
\arguments{
\item{Y}{A vector of outcomes}

\item{X}{A vector of predictors}

\item{order}{The order of the polynomial}

\item{space}{The values at which to predict}
}
\description{
This performs a sive estimation
}
