% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_transform.R
\name{transform_profile}
\alias{transform_profile}
\title{Transform the expression profile in `assayData` sample by sample}
\usage{
transform_profile(object, transform = c("identity", "log10", "log10p"))
}
\arguments{
\item{object, }{Object; a [`matrix`] or [`assayData-class`] or [`ExpressionSet-class`].}

\item{transform, }{Character; transformation to apply, the options inclulde:
* "identity", return the original data without any transformation.
* "log10", the transformation is `log10(object)`, and if the data contains
  zeros the transformation is `log10(1 + object)`.
* "log10p", the transformation is `log10(1 + object)`.}
}
\value{
A object matches the class of argument `object` with the transformed
  `assayData`.
}
\description{
Transform the taxa abundances in `assayData` sample by sample, which means
the counts of each sample will be transformed individually.
}
\examples{

}
\references{
https://github.com/yiluheihei/microbiomeMarker.
}
