% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsls.R
\name{tsls}
\alias{tsls}
\title{Two-stage least squares}
\usage{
tsls(Y, X, Z, W)
}
\arguments{
\item{Y}{The outcome variables}

\item{X}{The endogenous variables}

\item{Z}{The excluded instruments}

\item{W}{The included instruments or controls}
}
\description{
Estimates TSLS coefficients.
}
