% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsls.var.R
\name{tsls.var}
\alias{tsls.var}
\title{Heteroskedasticity robust tsls covariance matrix}
\usage{
tsls.var(Y, X, Z, W, beta)
}
\arguments{
\item{Y}{The Y data used in estimation}

\item{X}{The X data used in estimation}

\item{Z}{The Z used in estimation}

\item{W}{The W used in estimation}

\item{beta}{The vector of regression estimates}
}
\description{
Computes tsls covariance matrix of estimates
}
