% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15.parametric_test.R
\name{ANOVA_two}
\alias{ANOVA_two}
\title{Two-Way ANOVA Test}
\usage{
ANOVA_two(x, y, DNAID, GROUP1, GROUP2)
}
\arguments{
\item{x}{x with sampleID and group; sampleID connected to y}

\item{y}{y table rownames->taxonomy; colnames->sampleID}

\item{DNAID}{names of sampleID to connect x and y}

\item{GROUP1}{names of group information: factor1}

\item{GROUP2}{names of group information: factor2}
}
\value{
Returns a result of Two-Way ANOVA Test

type:     kind of data

Num:      number of group

P-value:  P by Two-Way ANOVA Test or post test pvalue

FDR:        adjusted by BH

mean+/-sd:   each group

median:     each group
}
\description{
Two-way ANOVA test is used to evaluate simultaneously the effect of
two grouping variables (A and B) on a response variable
}
\details{
07/18/2019
}
\examples{
result <- ANOVA_two(phen, prof, "SampleID", "Stage", "Group")
}
\author{
Hua Zou
}
