% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.var.R
\name{cluster.var}
\alias{cluster.var}
\title{Cluster corrected covariance matrix}
\usage{
cluster.var(Y, X, beta, groupvar, const = TRUE)
}
\arguments{
\item{Y}{The Y data used in estimation}

\item{X}{The X data used in estimation}

\item{beta}{The vector of linear regression estimates}

\item{groupvar}{A vector of group categorical variables assigning each
element of Y to a group}

\item{const}{If TRUE, adds a constant column to the data}
}
\description{
Computes the cluster corrected covariance matrix of estimates
}
