% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_filter.R
\name{filter_method}
\alias{filter_method}
\title{Filter samples or features in `assayData` by Occurrences}
\usage{
filter_method(Object, Cutoff=0.2, FilterType="Both")
}
\arguments{
\item{Cutoff, }{Numeric; the threshold for filtering (default: Cutoff=0.2).}

\item{FilterType, }{Character; the type of filtering data ("Both", "Features", "Samples").}

\item{Object, }{Object; a [`matrix`] or [`assayData-class`] or [`ExpressionSet-class`].}
}
\value{
A filtered `object` with the Cutoff.
}
\description{
Filter samples or features in `assayData` by Occurrences,
which means the samples or features will be discard if they could not pass the cutoff.
}
\details{
12/6/2021 Guangzhou China
}
\examples{
\donttest{
   data("ExprSet_species")
   filter_method(ExprSet_species, Cutoff=0.2, FilterType="Both")
}

}
\author{
Hua Zou
}
