% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_ExpressionSet.R
\name{get_ExprSet}
\alias{get_ExprSet}
\title{Building ExpressionSet Object}
\usage{
get_ExprSet(profile=Profile,
                  metadata=Metadata,
                  feature=Feature,
                  name="",
                  lab="",
                  contact="",
                  title="",
                  abstract="",
                  url="",
                  notes="")
}
\arguments{
\item{profile, }{Numeric matrix; (Required)a Matrix of expression data, whose Row is FeatureID and Column is SampleID.}

\item{metadata, }{Data.frame; (Required)a dataframe. of Metadata(1st column must be "SampleID"), containing Group information and also environmental factors(biological factors).}

\item{feature, }{Data.frame; the feature of Profile(default: Feature=NULL).}

\item{name, }{Character; name of experimentData, inheriting from ExpressionSet(default: name="").}

\item{lab, }{Character; lab of experimentData(default: lab="").}

\item{contact, }{Character; contact of experimentData(default: contact="").}

\item{title, }{Character; title of experimentData(default: title="").}

\item{abstract, }{Character; abstract of experimentData(default: abstract="").}

\item{url, }{Character; url of experimentData(default: url="").}

\item{notes, }{Character; notes of experimentData(default: notes="").}
}
\value{
an ExpressionSet Object
}
\description{
ExpressionSet object comprising not only expressiondata but also metadata is very convenient to perform data analysis.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

\donttest{
library(dplyr)
Profile <- data.table::fread(system.file("extdata", "Species_relative_abundance.tsv", package="MyRtools"))  \%>\% tibble::column_to_rownames("V1")
Metadata <- read.csv(system.file("extdata", "Metadata.csv", package="MyRtools"))
Feature <- read.csv(system.file("extdata", "Species_feature.csv", package="MyRtools")) \%>\% tibble::column_to_rownames("Species")

ExprSet <- get_ExprSet(profile=Profile,
                       metadata=Metadata,
                       feature=Feature,
                       name="Hua Zou",
                       lab="UCAS",
                       contact="zouhua1@outlook.com",
                       title="Experiment",
                       abstract="Profile",
                       url="www.zouhua.top",
                       notes="Expression")
}

}
\author{
Hua Zou
}
