% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet.R
\name{get_ExprSet}
\alias{get_ExprSet}
\title{Building ExpressionSet Object}
\usage{
get_ExprSet(Profile, Metadata, Feature=NULL)
}
\arguments{
\item{Profile, }{(Required)a Matrix of expression data, whose Row is FeatureID and Column is SampleID.}

\item{Metadata, }{(Required)a dataframe. of Metadata(1st column must be "SampleID"), containing Group information and also environmental factors(biological factors).}

\item{Feature, }{a dataframe of the feature of Profile.}

\item{Trim, }{filtering Feature and Sample with occurrence(default: FALSE).}

\item{Occ_Feature, }{the cutoff of Feature's trim(default: occurrence=0.2).}

\item{Each, }{filtering Features each group or whole data(default: Each("Group")=FALSE).}

\item{Occ_Sample, }{the cutoff of Sample's trim(default: occurrence=0.2).}
}
\value{
an ExpressionSet Object
}
\description{
ExpressionSet object comprising not only expressiondata but also metadata is very convenient to perform data analysis.
}
\details{
12/2/2021 Guangzhou China
}
\examples{

Profile <- data.table::fread(system.file("extdata", "Species_relative_abundance.tsv", package="MyRtools"))  \%>\% tibble::column_to_rownames("V1")
Metadata <- read.csv(system.file("extdata", "Metadata.csv", package="MyRtools"))
Feature <- read.csv(system.file("extdata", "Species_feature.csv", package="MyRtools")) \%>\% tibble::column_to_rownames("Species")

get_ExprSet(profile=Profile, metadata=Metadata, feature=Feature, trim=TRUE, occ_Feature=0.2, each=TRUE, occ_Sample=0.2)

}
\author{
Hua Zou
}
