% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_ExpressionSet.R
\name{get_processedExprSet}
\alias{get_processedExprSet}
\title{Building ExpressionSet Object with preprocess}
\usage{
get_processedExprSet(dataset=ExpressionSet,
        trim_cutoff=0.2,
        trim_type=c("identity", "both", "feature", "sample", "Group"),
        transform=c("identity", "log2", "log2p", "log10", "log10p"),
        imputation=c("identity", "knn"),
        normalization=c("identity", "TSS"))
}
\arguments{
\item{dataset, }{ExpressionSet; (Required) an Raw ExpressionSet object by `get_ExprSet`.}

\item{trim_cutoff, }{Numeric; the threshold for filtering (default: Cutoff=0.2).}

\item{trim_type, }{Data.frame; the feature of Profile.}

\item{normalization, }{Data.frame; the feature of Profile.}

\item{tranform, }{Data.frame; the feature of Profile.}

\item{inputation, }{Data.frame; the feature of Profile.}
}
\value{
an preprocessed ExpressionSet Object
}
\description{
precessing the ExpressionSet object by filtering, transforming or normalization
}
\details{
12/7/2021 Guangzhou China
}
\examples{

\donttest{
library(dplyr)
data("ExprSetRawRB")
ExprSet <- get_processedExprSet(dataset=ExprSetRawCount,
trim_cutoff=0.2,
trim_type="Group",
transform="log2",,
imputation=NULL,
normalization=NULL)
}

}
\author{
Hua Zou
}
