% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc0.var.R
\name{hc0.var}
\alias{hc0.var}
\title{Heteroskedasticity robust HC0 covariance matrix}
\usage{
hc0.var(Y, X, beta)
}
\arguments{
\item{Y}{The Y data used in estimation}

\item{X}{The X data used in estimation}

\item{beta}{The vector of linear regression estimates}
}
\description{
Computes HC0 covariance matrix of estimates
}
