% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.reg.R
\name{lin.reg}
\alias{lin.reg}
\title{Basic linear regression}
\usage{
lin.reg(Y, X)
}
\arguments{
\item{Y}{The outcome variables}

\item{X}{The explanatory variables}
}
\description{
Estimates linear regression coefficients. Standard errors should be computed
with a different function depending on various assumptions.
}
