% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Classes.R
\name{MyDataSet}
\alias{MyDataSet}
\title{Build MyDataSet-class objects}
\usage{
MyDataSet(Profile_table = NULL, Sample_data = NULL, Feature_table = NULL)
}
\arguments{
\item{Profile_table}{a [`marker_table-class`] object differtial analysis.}

\item{Sample_data}{character, method used to normalize the input `phyloseq`
object.}

\item{Feature_table}{character, method used for microbiome marker
identification.}

\item{...}{arguments}
}
\value{
a [`MyDataSet-class`] object.
}
\description{
This the constructor to build the [`MyDataSet-class`] object.
}
\examples{
\donttest{
library(dplyr)
Profile <- data.table::fread(system.file("extdata", "Species_relative_abundance.tsv", package="MyRtools"))  \%>\% tibble::column_to_rownames("V1")
Metadata <- read.csv(system.file("extdata", "Metadata.csv", package="MyRtools"))
Feature <- read.csv(system.file("extdata", "Species_feature.csv", package="MyRtools")) \%>\% tibble::column_to_rownames("Species")

PROF <- Profile_table(as.matrix(Profile), feature_are_row=TRUE)
META <- Sample_data(Metadata)
FEATURE <- Feature_table(Feature)

mydataset <- MyDataSet(Profile_table=PROF,
          Sample_data=META,
          Feature_table=FEATURE)
}

}
\seealso{
[phyloseq::phyloseq()]
}
