% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.parametric_test.R
\name{paired_ttest}
\alias{paired_ttest}
\title{Paired T-Test}
\usage{
paired_ttest(x, y, DNAID, GROUP, grp1=NULL, grp2=NULL)
}
\arguments{
\item{x}{x with sampleID, ID and group; sampleID connected to x}

\item{y}{y table rownames->taxonomy; colnames->sampleID}

\item{DNAID}{names of sampleID to connect x and y}

\item{PID}{id for paired test}

\item{GROUP}{names of group information, only contail two levels if grp1 or grp2 haven't been provided}

\item{grp1}{one of groups to be converted into 0}

\item{grp2}{one of groups to be converted into 1}
}
\value{
Returns a result of Paired T-Test

type:       kind of data

Block:      group information

Num:        number of group

P-value:    P by Paired T-Test

FDR:        adjusted by BH

Enrichment: directory by median or directory by mean

Occurence:  occurence of two groups

median:     both or each group

mean:       each group

FDR:        adjusted P value by BH

Odds Ratio:     95% Confidence interval
}
\description{
Tests for the difference between two variables from the same population (e.g., a pre- and posttest score)
}
\details{
07/18/2019   ShenZhen China
}
\examples{
result <- paired_ttest(phen, prof, "SampleID", "ID","Stage", "BASE", "WASH")
}
\author{
Hua Zou
}
