% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_PERMANOVA.R
\name{PERMANOVA}
\alias{PERMANOVA}
\title{Permutational Multivariate Analysis of Variance (PERMANOVA) Using Distance Matrices}
\usage{
PERMANOVA(dataset=ExpressionSet, Group_info="Group", Group_name=NULL, Distance="bray")
}
\arguments{
\item{Group_info, }{Character; design factor(default: "Group").}

\item{Group_name, }{Character; the group for filtering(default: NULL).}

\item{Distance, }{Character; Normalizing feature(default: Distance="bray").}

\item{Expression, }{ExpressionSet; (Required) A \code{ExpressionSet} object contains expression profile and
metadata including the continuous and categorical variables.}
}
\value{
PERMANOVA's results
}
\description{
The PERMANOVA is aimed to assess the association between the overall profile and metadata information.
It comprises permutation and ANOVA by using the F distribution to acquire the statistic.
}
\details{
12/4/2021 Guangzhou China
}
\examples{

\donttest{
data(ExprSet_species)

PERMANOVA_res <- PERMANOVA(dataset=ExprSet_species, Group_info="Group", Group_name=NULL, Distance="bray")
}

}
\author{
Hua Zou
}
