% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_filter.R
\name{run_filter}
\alias{run_filter}
\title{Filter samples or features in `assayData` by Occurrences}
\usage{
run_filter(object, object=0.2, filterType="Both")
}
\arguments{
\item{object, }{Object; a [`matrix`] or [`assayData-class`] or [`ExpressionSet-class`].}

\item{cutoff, }{Numeric; the threshold for filtering (default: Cutoff=0.2).}

\item{filterType, }{Character; transformation to apply, the options inclulde:
* "none", return the original data without any filter.
* "both", filter the features and samples of profile more than cutoff.
* "feature", filter the features of profile more than cutoff.
* "sample", filter the samples of profile more than cutoff.
* "Group", filter the features per Group, the samples of the profile more than cutoff, respectively.}
}
\value{
A filtered `object` with profile more than cutoff.
}
\description{
Filter samples or features in `assayData` by Occurrences,
which means the samples or features will be discard if they could not pass the cutoff.
}
\details{
12/6/2021 Guangzhou China
}
\examples{
\donttest{
   data("ExprSet_species")
   run_filter(ExprSet_species, object=0.2, filterType="both")
}

}
\author{
Hua Zou
}
