% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_normalization.R
\name{run_normalize}
\alias{run_normalize}
\title{Normalize the expression profile data}
\usage{
run_normalize(object, normalize=c("none", "TSS", "TMM", "RLE", "CSS", "CLR", "CPM", "Zscore"))
}
\arguments{
\item{object, }{Object; a [`matrix`] or [`assayData-class`] or [`ExpressionSet-class`].}

\item{normalize, }{Character; normalization to apply, the options inclulde:
* "none", return the original data without any transformation.
* "TSS": total sum scaling, also referred to as "relative abundance", the
    abundances were normalized by dividing the corresponding sample library
    size.
* "TMM": trimmed mean of m-values. First, a sample is chosen as reference.
    The scaling factor is then derived using a weighted trimmed mean over
    the differences of the log-transformed gene-count fold-change between
    the sample and the reference.
* "RLE", relative log expression, RLE uses a pseudo-reference calculated
    using the geometric mean of the gene-specific abundances over all
    samples. The scaling factors are then calculated as the median of the
    gene counts ratios between the samples and the reference.
* "CSS": cumulative sum scaling, calculates scaling factors as the
    cumulative sum of gene abundances up to a data-derived threshold.
* "CLR": centered log-ratio normalization.
* "CPM": pre-sample normalization of the sum of the values to 1e+06.
* "Zscore": convert the data into unit normal distribution(mean=0; sd=1).}
}
\value{
A object matches the class of argument `object` with the normalized profile.
}
\description{
Eliminating the bias such as sequencing depth, platform and other technical effects is vital for downstream data analysis.
}
\details{
12/7/2021 Guangzhou China
}
\examples{

}
\references{
https://github.com/yiluheihei/microbiomeMarker.
}
\author{
Hua Zou
}
