% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_transform.R
\name{transform_profile}
\alias{transform_profile}
\title{Transform the profile abundances in `Profile_table` sample by sample}
\usage{
transform_profile(object, transform = c("identity", "log10", "log10p"))
}
\arguments{
\item{object, }{Class; [`Profile_table-class`] or [`MyDataSet-class`].}

\item{transform, }{Character; transformation to apply, the options inclulde:
* "identity", return the original data without any transformation.
* "log10", the transformation is `log10(object)`, and if the data contains
  zeros the transformation is `log10(1 + object)`.
* "log10p", the transformation is `log10(1 + object)`.}
}
\value{
A object matches the class of argument `object` with the transformed
  `Profile_table`.
}
\description{
Transform the taxa abundances in `Profile_table` sample by sample, which means
the counts of each sample will be transformed individually.
}
\examples{
\donttest{
data(mydataset)
x1 <- transform_profile(mydataset)
head(Profile_table(x1), 10)
x2 <- transform_profile(mydataset, "log10")
head(Profile_table(x2), 10)
x3 <- transform_profile(mydataset, "log10p")
head(Profile_table(x3), 10)
}

}
\references{
https://github.com/yiluheihei/microbiomeMarker.
}
