% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{venn_plot}
\alias{venn_plot}
\title{plot a venn graphics}
\usage{
venn_plot(data, fillColor = NULL, fileName = NULL, main = NULL,
  fontSize = 1.5, fontColor = "white", saveType = c("tiff", "png"),
  Margin = 0.1)
}
\arguments{
\item{data}{a list for your group data}

\item{fillColor}{a vector for fill your venn plot default \code{fillColor} is provided}

\item{fileName}{a character for your outputfile name}

\item{main}{a character for the title in your venn plot default \code{main} is NULL}

\item{fontSize}{a number for your label size default \code{fontSize} is 1.5}

\item{fontColor}{a character for your label color default \code{fontColor} is white}

\item{saveType}{a character for your save file type png or tiff}

\item{Margin}{a number for your venn margin more Margin less size in your venn graphics}
}
\value{
a tiff or png file for your venn graphics
}
\description{
plot a venn graphics it wrap VennDiagram and make venn more simple
}
\note{
Since the \code{venn.diaram()} is not support pdf type file so your \code{fillName} not allow pdf type file
}
\examples{

oneName <- function() paste(sample(LETTERS,5,replace=TRUE),collapse="")
geneNames <- replicate(1000,oneName())
GroupA <- sample(geneNames,400,replace = F)
GroupB <- sample(geneNames,300,replace = F)
GroupC <- sample(geneNames,200,replace = F)
GroupD <- sample(geneNames,500,replace = F)
venn_plot(data=list(A=GroupA,B=GroupB,C=GroupC,D=GroupD),fileName='test_venn_plot.tiff')
}
\author{
chencheng <chencheng@onmath.cn>
}
